refactor(bench): split comparison harness modules
