+++ crates/ferrocat-bench/src/compare.rs
@@ -1,3 +1,19 @@
+mod adapters;
+mod environment;
+mod operations;
+mod profile;
+mod regression;
+mod report;
+mod semantic_digest;
+
+use adapters::*;
+use environment::*;
+use operations::*;
+use profile::*;
+use regression::*;
+use report::*;
+use semantic_digest::*;
+
 use std::borrow::Cow;
 use std::collections::{BTreeMap, BTreeSet};
 use std::env;
@@ -466,3767 +482,4 @@ fn parse_positive_f64(label: &str, value: &str) -> Result<f64, String> {
     Ok(parsed)
 }
 
-#[derive(Debug)]
-struct BenchmarkRegressionReport {
-    profile: String,
-    max_regression_percent: f64,
-    passed: Vec<BenchmarkRegressionScenario>,
-    failures: Vec<BenchmarkRegressionScenario>,
-    skipped_noisy: Vec<BenchmarkRegressionScenario>,
-    skipped_semantics_changed: Vec<BenchmarkSemanticsChangedScenario>,
-    missing_baseline: Vec<String>,
-    missing_current: Vec<String>,
-}
-
-impl BenchmarkRegressionReport {
-    fn has_failures(&self) -> bool {
-        !self.failures.is_empty() || !self.missing_current.is_empty()
-    }
-
-    fn render(&self) -> String {
-        let mut out = String::new();
-        out.push_str("Benchmark regression check\n");
-        out.push_str(&format!("profile: {}\n", self.profile));
-        out.push_str(&format!(
-            "max-regression-percent: {:.2}\n",
-            self.max_regression_percent
-        ));
-        out.push_str(&format!(
-            "passed: {} failed: {} skipped-noisy: {} skipped-semantics-changed: {} missing-baseline: {} missing-current: {}\n",
-            self.passed.len(),
-            self.failures.len(),
-            self.skipped_noisy.len(),
-            self.skipped_semantics_changed.len(),
-            self.missing_baseline.len(),
-            self.missing_current.len()
-        ));
-
-        for scenario in &self.failures {
-            out.push_str(&format!(
-                "FAIL {}: +{:.2}% per-iteration time ({} -> {})\n",
-                scenario.id,
-                scenario.regression_percent,
-                format_per_iteration_ns(scenario.baseline_per_iteration_ns),
-                format_per_iteration_ns(scenario.current_per_iteration_ns)
-            ));
-        }
-        for scenario in &self.skipped_noisy {
-            out.push_str(&format!(
-                "SKIP noisy {}: {:+.2}% per-iteration time (baseline noisy={} current noisy={})\n",
-                scenario.id,
-                scenario.regression_percent,
-                scenario.baseline_noisy,
-                scenario.current_noisy
-            ));
-        }
-        for scenario in &self.skipped_semantics_changed {
-            out.push_str(&format!(
-                "SKIP semantics-changed {}: baseline digest {} -> current digest {}\n",
-                scenario.id, scenario.baseline_semantic_digest, scenario.current_semantic_digest
-            ));
-        }
-        for id in &self.missing_baseline {
-            out.push_str(&format!("SKIP missing-baseline {id}\n"));
-        }
-        for id in &self.missing_current {
-            out.push_str(&format!("FAIL missing-current {id}\n"));
-        }
-        if self.has_failures() {
-            out.push_str("result: FAIL\n");
-        } else {
-            out.push_str("result: PASS\n");
-        }
-        out
-    }
-}
-
-#[derive(Debug)]
-struct BenchmarkRegressionScenario {
-    id: String,
-    baseline_per_iteration_ns: f64,
-    current_per_iteration_ns: f64,
-    regression_percent: f64,
-    baseline_noisy: bool,
-    current_noisy: bool,
-}
-
-#[derive(Debug)]
-struct BenchmarkSemanticsChangedScenario {
-    id: String,
-    baseline_semantic_digest: String,
-    current_semantic_digest: String,
-}
-
-fn compare_regression_reports(
-    baseline: &CompareReport,
-    current: &CompareReport,
-    max_regression_percent: f64,
-) -> Result<BenchmarkRegressionReport, String> {
-    if baseline.profile != current.profile {
-        return Err(format!(
-            "cannot compare benchmark profiles {} and {}",
-            baseline.profile, current.profile
-        ));
-    }
-
-    let baseline_by_id = baseline
-        .scenarios
-        .iter()
-        .map(|scenario| (scenario.id.as_str(), scenario))
-        .collect::<BTreeMap<_, _>>();
-    let current_by_id = current
-        .scenarios
-        .iter()
-        .map(|scenario| (scenario.id.as_str(), scenario))
-        .collect::<BTreeMap<_, _>>();
-
-    let mut passed = Vec::new();
-    let mut failures = Vec::new();
-    let mut skipped_noisy = Vec::new();
-    let mut skipped_semantics_changed = Vec::new();
-    let mut missing_baseline = Vec::new();
-    let mut missing_current = Vec::new();
-
-    for scenario in &current.scenarios {
-        let Some(baseline_scenario) = baseline_by_id.get(scenario.id.as_str()) else {
-            missing_baseline.push(scenario.id.clone());
-            continue;
-        };
-        if baseline_scenario.semantic_digest != scenario.semantic_digest {
-            skipped_semantics_changed.push(BenchmarkSemanticsChangedScenario {
-                id: scenario.id.clone(),
-                baseline_semantic_digest: baseline_scenario.semantic_digest.clone(),
-                current_semantic_digest: scenario.semantic_digest.clone(),
-            });
-            continue;
-        }
-        let regression = regression_scenario(baseline_scenario, scenario)?;
-        if regression.baseline_noisy || regression.current_noisy {
-            skipped_noisy.push(regression);
-        } else if regression.regression_percent > max_regression_percent {
-            failures.push(regression);
-        } else {
-            passed.push(regression);
-        }
-    }
-
-    for scenario in &baseline.scenarios {
-        if !current_by_id.contains_key(scenario.id.as_str()) {
-            missing_current.push(scenario.id.clone());
-        }
-    }
-
-    if passed.is_empty()
-        && failures.is_empty()
-        && skipped_noisy.is_empty()
-        && skipped_semantics_changed.is_empty()
-        && missing_baseline.is_empty()
-        && missing_current.is_empty()
-    {
-        return Err("benchmark reports have no comparable scenarios".to_owned());
-    }
-
-    Ok(BenchmarkRegressionReport {
-        profile: current.profile.clone(),
-        max_regression_percent,
-        passed,
-        failures,
-        skipped_noisy,
-        skipped_semantics_changed,
-        missing_baseline,
-        missing_current,
-    })
-}
-
-fn regression_scenario(
-    baseline: &ScenarioReport,
-    current: &ScenarioReport,
-) -> Result<BenchmarkRegressionScenario, String> {
-    // Compare per-iteration time, not raw sample time. The harness calibrates
-    // iterations_per_sample independently per run, so a faster build can pick
-    // more iterations per sample and end up with a *larger* median_elapsed_ns
-    // even though each operation got quicker. Normalizing by iterations keeps
-    // the check measuring the work, not the calibration.
-    let baseline_per_iteration = per_iteration_nanos(baseline)?;
-    let current_per_iteration = per_iteration_nanos(current)?;
-
-    let regression_percent = ((current_per_iteration / baseline_per_iteration) - 1.0) * 100.0;
-
-    Ok(BenchmarkRegressionScenario {
-        id: current.id.clone(),
-        baseline_per_iteration_ns: baseline_per_iteration,
-        current_per_iteration_ns: current_per_iteration,
-        regression_percent,
-        baseline_noisy: baseline.statistics.noisy,
-        current_noisy: current.statistics.noisy,
-    })
-}
-
-fn per_iteration_nanos(report: &ScenarioReport) -> Result<f64, String> {
-    if report.iterations_per_sample == 0 {
-        return Err(format!(
-            "scenario {} has zero iterations_per_sample",
-            report.id
-        ));
-    }
-    if report.statistics.median_elapsed_ns == 0 {
-        return Err(format!(
-            "scenario {} has zero median elapsed time",
-            report.id
-        ));
-    }
-    Ok(f64_from_u128(report.statistics.median_elapsed_ns) / report.iterations_per_sample as f64)
-}
-
-fn format_per_iteration_ns(value: f64) -> String {
-    format!("{:.4} ms", value / 1_000_000.0)
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct BenchmarkProfile {
-    name: String,
-    #[serde(default = "default_minimum_sample_millis")]
-    minimum_sample_millis: u64,
-    scenarios: Vec<BenchmarkScenario>,
-}
-
-impl BenchmarkProfile {
-    fn load(workspace: &Path, profile_name: &str) -> Result<Self, String> {
-        let path = workspace
-            .join("benchmark")
-            .join("profiles")
-            .join(format!("{profile_name}.json"));
-        let content = fs::read_to_string(&path).map_err(|error| {
-            format!(
-                "failed to read benchmark profile {}: {error}",
-                path.display()
-            )
-        })?;
-        let profile: Self = serde_json::from_str(&content).map_err(|error| {
-            format!(
-                "failed to parse benchmark profile {}: {error}",
-                path.display()
-            )
-        })?;
-        if profile.name != profile_name {
-            return Err(format!(
-                "benchmark profile {} declares mismatched name {}",
-                path.display(),
-                profile.name
-            ));
-        }
-        if profile.scenarios.is_empty() {
-            return Err(format!(
-                "benchmark profile {} defines no scenarios",
-                path.display()
-            ));
-        }
-        Ok(profile)
-    }
-
-    fn tool_requirement(&self) -> ToolRequirement {
-        if self
-            .scenarios
-            .iter()
-            .any(|scenario| !scenario.implementation.starts_with("ferrocat-"))
-        {
-            ToolRequirement::External
-        } else {
-            ToolRequirement::RustOnly
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ToolRequirement {
-    RustOnly,
-    External,
-}
-
-const fn default_minimum_sample_millis() -> u64 {
-    DEFAULT_MIN_SAMPLE_MILLIS
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct BenchmarkScenario {
-    id: String,
-    comparison_group: String,
-    workload: String,
-    operation: String,
-    fixture: String,
-    implementation: String,
-    warmup_runs: usize,
-    measured_runs: usize,
-    minimum_sample_millis: Option<u64>,
-}
-
-#[derive(Debug)]
-struct PreparedScenario {
-    operation: String,
-    fixture: String,
-    tempdir: TempDir,
-    po_input_path: Option<PathBuf>,
-    icu_messages_path: Option<PathBuf>,
-    existing_po_path: Option<PathBuf>,
-    pot_path: Option<PathBuf>,
-    po_content: Option<String>,
-    catalog_fcl_content: Option<String>,
-    po_file: Option<PoFile>,
-    merge_fixture: Option<OwnedMergeFixture>,
-    icu_messages: Option<Vec<String>>,
-}
-
-#[derive(Debug)]
-struct CliBaselineScenario {
-    label: String,
-    prepared: PreparedScenario,
-}
-
-#[derive(Debug)]
-struct ScenarioExecutionPlan {
-    scenario: BenchmarkScenario,
-    tool_version: String,
-    validated_digest: String,
-    iterations: usize,
-    cli_baseline: Option<CliBaselineScenario>,
-}
-
-impl PreparedScenario {
-    #[expect(
-        clippy::too_many_lines,
-        reason = "Scenario preparation is intentionally centralized so fixture setup remains deterministic."
-    )]
-    fn prepare(workspace: &Path, scenarios: &[BenchmarkScenario]) -> Result<Self, String> {
-        let Some(first) = scenarios.first() else {
-            return Err("cannot prepare empty benchmark scenario group".to_owned());
-        };
-        for scenario in scenarios {
-            if scenario.operation != first.operation || scenario.fixture != first.fixture {
-                return Err(format!(
-                    "comparison group {} mixes incompatible operations or fixtures",
-                    first.comparison_group
-                ));
-            }
-        }
-
-        let tempdir = tempfile::Builder::new()
-            .prefix("ferrocat-compare-")
-            .tempdir_in(workspace.join("target"))
-            .map_err(|error| format!("failed to create compare tempdir: {error}"))?;
-
-        match first.operation.as_str() {
-            "parse" | "stringify" => {
-                let fixture = load_fixture(&first.fixture)?;
-                let input_path = tempdir.path().join("input.po");
-                fs::write(&input_path, fixture.content()).map_err(|error| {
-                    format!(
-                        "failed to write fixture input {}: {error}",
-                        input_path.display()
-                    )
-                })?;
-                let po_file = parse_po(fixture.content())
-                    .map_err(|error| format!("failed to parse fixture: {error}"))?;
-                Ok(Self {
-                    operation: first.operation.clone(),
-                    fixture: first.fixture.clone(),
-                    tempdir,
-                    po_input_path: Some(input_path),
-                    icu_messages_path: None,
-                    existing_po_path: None,
-                    pot_path: None,
-                    po_content: Some(fixture.content().to_owned()),
-                    catalog_fcl_content: None,
-                    po_file: Some(po_file),
-                    merge_fixture: None,
-                    icu_messages: None,
-                })
-            }
-            "parse-catalog" => {
-                let fixture = load_fixture(&first.fixture)?;
-                let po_input_path = tempdir.path().join("catalog-input.po");
-                fs::write(&po_input_path, fixture.content()).map_err(|error| {
-                    format!(
-                        "failed to write catalog fixture input {}: {error}",
-                        po_input_path.display()
-                    )
-                })?;
-                let po_content = fixture.content().to_owned();
-                // Render an equivalent FCL catalog so the storage-format comparison
-                // can parse the same logical catalog as both PO and FCL.
-                let mut options =
-                    ParseCatalogOptions::new(&po_content, "en").with_mode(CatalogMode::IcuPo);
-                let fixture_locale = fixture_locale(&first.fixture);
-                if let Some(locale) = fixture_locale.as_deref() {
-                    options = options.with_locale(locale);
-                }
-                let parsed = parse_catalog(options)
-                    .map_err(|error| format!("failed to parse catalog fixture: {error}"))?;
-                let fcl_content = crate::render_fcl_catalog(&parsed);
-                Ok(Self {
-                    operation: first.operation.clone(),
-                    fixture: first.fixture.clone(),
-                    tempdir,
-                    po_input_path: Some(po_input_path),
-                    icu_messages_path: None,
-                    existing_po_path: None,
-                    pot_path: None,
-                    po_content: Some(po_content),
-                    catalog_fcl_content: Some(fcl_content),
-                    po_file: None,
-                    merge_fixture: None,
-                    icu_messages: None,
-                })
-            }
-            "merge" | "update-catalog" => {
-                let fixture = load_merge_fixture(&first.fixture)?;
-                let existing_po_path = tempdir.path().join("existing.po");
-                fs::write(&existing_po_path, fixture.existing_po()).map_err(|error| {
-                    format!(
-                        "failed to write merge fixture input {}: {error}",
-                        existing_po_path.display()
-                    )
-                })?;
-                let pot_path = if first.operation == "merge"
-                    || scenarios
-                        .iter()
-                        .any(|scenario| !scenario.implementation.starts_with("ferrocat-"))
-                {
-                    let pot_path = tempdir.path().join("template.pot");
-                    let pot = build_merge_pot(&fixture);
-                    fs::write(&pot_path, pot).map_err(|error| {
-                        format!(
-                            "failed to write merge template {}: {error}",
-                            pot_path.display()
-                        )
-                    })?;
-                    Some(pot_path)
-                } else {
-                    None
-                };
-                Ok(Self {
-                    operation: first.operation.clone(),
-                    fixture: first.fixture.clone(),
-                    tempdir,
-                    po_input_path: None,
-                    icu_messages_path: None,
-                    existing_po_path: Some(existing_po_path),
-                    pot_path,
-                    po_content: None,
-                    catalog_fcl_content: None,
-                    po_file: None,
-                    merge_fixture: Some(OwnedMergeFixture::from_fixture(&fixture)),
-                    icu_messages: None,
-                })
-            }
-            "parse-icu" => {
-                let fixture = load_icu_fixture(&first.fixture)?;
-                let messages_path = tempdir.path().join("messages.json");
-                let messages = fixture.messages().to_vec();
-                let serialized = serde_json::to_string_pretty(&messages)
-                    .map_err(|error| format!("failed to serialize ICU messages: {error}"))?;
-                fs::write(&messages_path, serialized).map_err(|error| {
-                    format!(
-                        "failed to write ICU fixture input {}: {error}",
-                        messages_path.display()
-                    )
-                })?;
-                Ok(Self {
-                    operation: first.operation.clone(),
-                    fixture: first.fixture.clone(),
-                    tempdir,
-                    po_input_path: None,
-                    icu_messages_path: Some(messages_path),
-                    existing_po_path: None,
-                    pot_path: None,
-                    po_content: None,
-                    catalog_fcl_content: None,
-                    po_file: None,
-                    merge_fixture: None,
-                    icu_messages: Some(messages),
-                })
-            }
-            other => Err(format!("unsupported benchmark operation: {other}")),
-        }
-    }
-
-    fn prepare_cli_baseline(
-        workspace: &Path,
-        scenario: &BenchmarkScenario,
-    ) -> Result<Option<CliBaselineScenario>, String> {
-        match scenario.implementation.as_str() {
-            "msgcat" => {
-                let tempdir = tempfile::Builder::new()
-                    .prefix("ferrocat-cli-baseline-")
-                    .tempdir_in(workspace.join("target"))
-                    .map_err(|error| format!("failed to create cli baseline tempdir: {error}"))?;
-                let locale = fixture_locale(&scenario.fixture);
-                let content = build_cli_baseline_po(locale.as_deref());
-                let input_path = tempdir.path().join("baseline.po");
-                fs::write(&input_path, &content).map_err(|error| {
-                    format!(
-                        "failed to write cli baseline input {}: {error}",
-                        input_path.display()
-                    )
-                })?;
-                Ok(Some(CliBaselineScenario {
-                    label: format!("header-only-po:{}", locale.as_deref().unwrap_or("default")),
-                    prepared: Self {
-                        operation: scenario.operation.clone(),
-                        fixture: format!("cli-baseline-msgcat-{}", scenario.fixture),
-                        tempdir,
-                        po_input_path: Some(input_path),
-                        icu_messages_path: None,
-                        existing_po_path: None,
-                        pot_path: None,
-                        po_content: Some(content),
-                        catalog_fcl_content: None,
-                        po_file: None,
-                        merge_fixture: None,
-                        icu_messages: None,
-                    },
-                }))
-            }
-            "msgmerge" => {
-                let tempdir = tempfile::Builder::new()
-                    .prefix("ferrocat-cli-baseline-")
-                    .tempdir_in(workspace.join("target"))
-                    .map_err(|error| format!("failed to create cli baseline tempdir: {error}"))?;
-                let locale = fixture_locale(&scenario.fixture);
-                let existing = build_cli_baseline_po(locale.as_deref());
-                let pot = build_cli_baseline_pot();
-                let existing_po_path = tempdir.path().join("baseline-existing.po");
-                let pot_path = tempdir.path().join("baseline-template.pot");
-                fs::write(&existing_po_path, &existing).map_err(|error| {
-                    format!(
-                        "failed to write cli baseline existing {}: {error}",
-                        existing_po_path.display()
-                    )
-                })?;
-                fs::write(&pot_path, &pot).map_err(|error| {
-                    format!(
-                        "failed to write cli baseline template {}: {error}",
-                        pot_path.display()
-                    )
-                })?;
-                Ok(Some(CliBaselineScenario {
-                    label: format!(
-                        "header-only-merge:{}",
-                        locale.as_deref().unwrap_or("default")
-                    ),
-                    prepared: Self {
-                        operation: scenario.operation.clone(),
-                        fixture: format!("cli-baseline-msgmerge-{}", scenario.fixture),
-                        tempdir,
-                        po_input_path: None,
-                        icu_messages_path: None,
-                        existing_po_path: Some(existing_po_path),
-                        pot_path: Some(pot_path),
-                        po_content: None,
-                        catalog_fcl_content: None,
-                        po_file: None,
-                        merge_fixture: None,
-                        icu_messages: None,
-                    },
-                }))
-            }
-            _ => Ok(None),
-        }
-    }
-
-    fn validate(&self, result: &ExecutionResult) -> Result<String, String> {
-        let digest = match self.operation.as_str() {
-            "parse" => match result.artifact.as_ref() {
-                Some(ExecutionArtifact::PoSummary(summary)) => digest_summary(summary)?,
-                _ => {
-                    return Err(format!(
-                        "scenario {} expected PO summary artifact",
-                        self.fixture
-                    ));
-                }
-            },
-            "parse-catalog" => match result.artifact.as_ref() {
-                Some(ExecutionArtifact::CatalogSummary(summary)) => digest_summary(summary)?,
-                _ => {
-                    return Err(format!(
-                        "scenario {} expected catalog summary artifact",
-                        self.fixture
-                    ));
-                }
-            },
-            "stringify" | "merge" | "update-catalog" => {
-                let content = match result.artifact.as_ref() {
-                    Some(ExecutionArtifact::RenderedPo(content)) => Cow::Borrowed(content.as_str()),
-                    Some(ExecutionArtifact::RenderedPoPath(path)) => {
-                        Cow::Owned(fs::read_to_string(path).map_err(|error| {
-                            format!(
-                                "failed to read rendered PO output {}: {error}",
-                                path.display()
-                            )
-                        })?)
-                    }
-                    _ => {
-                        return Err(format!(
-                            "scenario {} expected rendered PO artifact for {}",
-                            self.fixture, self.operation
-                        ));
-                    }
-                };
-                let parsed = parse_po(&content)
-                    .map_err(|error| format!("rendered output did not parse as PO: {error}"))?;
-                digest_summary(&PoSemanticSummary::from_po_file(&parsed))?
-            }
-            "parse-icu" => match result.artifact.as_ref() {
-                Some(ExecutionArtifact::IcuSummary(summary)) => digest_summary(summary)?,
-                _ => {
-                    return Err(format!(
-                        "scenario {} expected ICU summary artifact",
-                        self.fixture
-                    ));
-                }
-            },
-            other => return Err(format!("unsupported validation operation: {other}")),
-        };
-        Ok(digest)
-    }
-
-    fn run_internal_parse(
-        &self,
-        iterations: usize,
-        borrowed: bool,
-    ) -> Result<ExecutionResult, String> {
-        let input = self
-            .po_content
-            .as_deref()
-            .ok_or_else(|| "internal parse requires PO content".to_owned())?;
-        // Time only the parse; build the digest summary once outside the loop so
-        // the measured path is parsing alone, matching the external adapters
-        // (which also reparse for the summary outside the timed loop).
-        let (elapsed, summary) = if borrowed {
-            let mut last = None;
-            let start = Instant::now();
-            for _ in 0..iterations {
-                last = Some(
-                    parse_po_borrowed(input)
-                        .map_err(|error| format!("borrowed parse failed: {error}"))?,
-                );
-            }
-            let elapsed = start.elapsed();
-            let parsed = last.ok_or_else(|| "internal parse produced no result".to_owned())?;
-            (elapsed, PoSemanticSummary::from_borrowed_po_file(&parsed))
-        } else {
-            let mut last = None;
-            let start = Instant::now();
-            for _ in 0..iterations {
-                last =
-                    Some(parse_po(input).map_err(|error| format!("owned parse failed: {error}"))?);
-            }
-            let elapsed = start.elapsed();
-            let parsed = last.ok_or_else(|| "internal parse produced no result".to_owned())?;
-            (elapsed, PoSemanticSummary::from_po_file(&parsed))
-        };
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: (input.len() * iterations) as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: Some(ExecutionArtifact::PoSummary(summary)),
-        })
-    }
-
-    fn run_internal_parse_catalog(&self, iterations: usize) -> Result<ExecutionResult, String> {
-        let content = self
-            .po_content
-            .as_deref()
-            .ok_or_else(|| "internal parse-catalog-po requires PO content".to_owned())?;
-        self.run_internal_parse_catalog_with(iterations, content, CatalogMode::IcuPo)
-    }
-
-    fn run_internal_parse_catalog_fcl(&self, iterations: usize) -> Result<ExecutionResult, String> {
-        let content = self
-            .catalog_fcl_content
-            .as_deref()
-            .ok_or_else(|| "internal parse-catalog-fcl requires FCL content".to_owned())?;
-        self.run_internal_parse_catalog_with(iterations, content, CatalogMode::IcuFcl)
-    }
-
-    fn run_internal_parse_catalog_with(
-        &self,
-        iterations: usize,
-        content: &str,
-        mode: CatalogMode,
-    ) -> Result<ExecutionResult, String> {
-        let locale = fixture_locale(&self.fixture);
-        let bytes_per_iteration = content.len();
-
-        // Time only parse_catalog; build the digest summary once outside the loop.
-        let mut last_parsed = None;
-        let start = Instant::now();
-        for _ in 0..iterations {
-            let mut options = ParseCatalogOptions::new(content, "en").with_mode(mode);
-            if let Some(locale) = locale.as_deref() {
-                options = options.with_locale(locale);
-            }
-            last_parsed = Some(
-                parse_catalog(options).map_err(|error| format!("parse_catalog failed: {error}"))?,
-            );
-        }
-        let elapsed = start.elapsed();
-        let parsed =
-            last_parsed.ok_or_else(|| "internal parse-catalog produced no result".to_owned())?;
-        let summary = CatalogSemanticSummary::from_parsed_catalog(parsed)?;
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: (bytes_per_iteration * iterations) as u64,
-            items_processed: None,
-            messages_processed: Some((summary.messages.len() * iterations) as u64),
-            artifact: Some(ExecutionArtifact::CatalogSummary(summary)),
-        })
-    }
-
-    fn run_internal_stringify(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let po_file = self
-            .po_file
-            .as_ref()
-            .ok_or_else(|| "internal stringify requires parsed PO file".to_owned())?;
-        let mut last_rendered = None;
-        let start = Instant::now();
-        let mut bytes_processed = 0usize;
-        for _ in 0..iterations {
-            let rendered = stringify_po(po_file, &SerializeOptions::default());
-            bytes_processed += rendered.len();
-            last_rendered = Some(rendered);
-        }
-        let elapsed = start.elapsed();
-        let rendered = last_rendered
-            .ok_or_else(|| "internal stringify produced no rendered content".to_owned())?;
-        let summary = {
-            let parsed = parse_po(&rendered).map_err(|error| {
-                format!("stringify validation parse failed for rendered output: {error}")
-            })?;
-            PoSemanticSummary::from_po_file(&parsed)
-        };
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: bytes_processed as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: capture_artifacts.then_some(ExecutionArtifact::RenderedPo(rendered)),
-        })
-    }
-
-    fn run_internal_merge(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let fixture = self
-            .merge_fixture
-            .as_ref()
-            .ok_or_else(|| "internal merge requires merge fixture".to_owned())?;
-        let mut last_rendered = None;
-        let start = Instant::now();
-        let mut bytes_processed = 0usize;
-        for _ in 0..iterations {
-            // Parse the same template.pot the competitors read, then merge.
-            // This keeps the catalog-update comparison strictly file-to-file
-            // instead of handing ferrocat pre-structured messages.
-            let template = parse_po(&fixture.template_pot)
-                .map_err(|error| format!("merge template parse failed: {error}"))?;
-            let extracted: Vec<MergeMessageInput<'_>> = template
-                .items
-                .iter()
-                .map(|item| MergeMessageInput {
-                    msgctxt: item.msgctxt.as_deref().map(Cow::Borrowed),
-                    msgid: Cow::Borrowed(item.msgid.as_str()),
-                    msgid_plural: item.msgid_plural.as_deref().map(Cow::Borrowed),
-                    references: item
-                        .references
-                        .iter()
-                        .map(|value| Cow::Borrowed(value.as_str()))
-                        .collect(),
-                    extracted_comments: item
-                        .extracted_comments
-                        .iter()
-                        .map(|value| Cow::Borrowed(value.as_str()))
-                        .collect(),
-                    flags: item
-                        .flags
-                        .iter()
-                        .map(|value| Cow::Borrowed(value.as_str()))
-                        .collect(),
-                })
-                .collect();
-            let rendered = merge_catalog(&fixture.existing_po, &extracted)
-                .map_err(|error| format!("merge_catalog failed: {error}"))?;
-            bytes_processed += rendered.len();
-            last_rendered = Some(rendered);
-        }
-        let elapsed = start.elapsed();
-        let rendered = last_rendered
-            .ok_or_else(|| "internal merge produced no rendered content".to_owned())?;
-        let summary = {
-            let parsed = parse_po(&rendered)
-                .map_err(|error| format!("merge output did not parse: {error}"))?;
-            PoSemanticSummary::from_po_file(&parsed)
-        };
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: bytes_processed as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: capture_artifacts.then_some(ExecutionArtifact::RenderedPo(rendered)),
-        })
-    }
-
-    fn run_internal_update_catalog(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let fixture = self
-            .merge_fixture
-            .as_ref()
-            .ok_or_else(|| "internal update-catalog requires merge fixture".to_owned())?;
-        let mut last_rendered = None;
-        let start = Instant::now();
-        let mut bytes_processed = 0usize;
-        let locale = fixture_locale(&self.fixture);
-        let mode = fixture_catalog_mode(&self.fixture);
-        for _ in 0..iterations {
-            let mut options = UpdateCatalogOptions::new("en", fixture.api_messages.clone())
-                .with_mode(mode)
-                .with_existing(fixture.existing_po.as_str());
-            if let Some(locale) = locale.as_deref() {
-                options = options.with_locale(locale);
-            }
-            let updated = update_catalog(options)
-                .map_err(|error| format!("update_catalog failed: {error}"))?;
-            bytes_processed += updated.content.len();
-            last_rendered = Some(updated.content);
-        }
-        let elapsed = start.elapsed();
-        let rendered = last_rendered
-            .ok_or_else(|| "internal update_catalog produced no rendered content".to_owned())?;
-        let summary = {
-            let parsed = parse_po(&rendered)
-                .map_err(|error| format!("update_catalog output did not parse: {error}"))?;
-            PoSemanticSummary::from_po_file(&parsed)
-        };
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: bytes_processed as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: capture_artifacts.then_some(ExecutionArtifact::RenderedPo(rendered)),
-        })
-    }
-
-    fn run_internal_update_catalog_file(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let fixture = self
-            .merge_fixture
-            .as_ref()
-            .ok_or_else(|| "internal update-catalog requires merge fixture".to_owned())?;
-        let mut last_rendered = None;
-        let start = Instant::now();
-        let mut bytes_processed = 0usize;
-        let locale = fixture_locale(&self.fixture);
-        let mode = fixture_catalog_mode(&self.fixture);
-        for _ in 0..iterations {
-            // Parse the same template.pot the competitors read and build the
-            // extraction input from it inside the timed loop. This keeps the
-            // catalog-update comparison strictly file-to-file instead of
-            // handing ferrocat pre-structured messages.
-            let template = parse_po(&fixture.template_pot)
-                .map_err(|error| format!("update template parse failed: {error}"))?;
-            let messages = extracted_messages_from_template(&template);
-            let mut options = UpdateCatalogOptions::new("en", messages)
-                .with_mode(mode)
-                .with_existing(fixture.existing_po.as_str());
-            if let Some(locale) = locale.as_deref() {
-                options = options.with_locale(locale);
-            }
-            let updated = update_catalog(options)
-                .map_err(|error| format!("update_catalog failed: {error}"))?;
-            bytes_processed += updated.content.len();
-            last_rendered = Some(updated.content);
-        }
-        let elapsed = start.elapsed();
-        let rendered = last_rendered
-            .ok_or_else(|| "internal update_catalog produced no rendered content".to_owned())?;
-        let summary = {
-            let parsed = parse_po(&rendered)
-                .map_err(|error| format!("update_catalog output did not parse: {error}"))?;
-            PoSemanticSummary::from_po_file(&parsed)
-        };
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: bytes_processed as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: capture_artifacts.then_some(ExecutionArtifact::RenderedPo(rendered)),
-        })
-    }
-
-    fn run_internal_parse_icu(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let messages = self
-            .icu_messages
-            .as_ref()
-            .ok_or_else(|| "internal parse-icu requires ICU messages".to_owned())?;
-        let total_bytes = messages.iter().map(String::len).sum::<usize>();
-        // Time only the parse; summarize the parsed messages once outside the loop.
-        let mut last_parsed = None;
-        let start = Instant::now();
-        for _ in 0..iterations {
-            let mut parsed = Vec::with_capacity(messages.len());
-            for message in messages {
-                parsed.push(
-                    parse_icu(message).map_err(|error| {
-                        format!("failed to parse ICU benchmark message: {error}")
-                    })?,
-                );
-            }
-            last_parsed = Some(parsed);
-        }
-        let elapsed = start.elapsed();
-        let parsed =
-            last_parsed.ok_or_else(|| "internal parse-icu produced no result".to_owned())?;
-        let summary = IcuFixtureSummary {
-            messages: parsed.iter().map(IcuMessageSummary::from_message).collect(),
-        };
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: INTERNAL_TOOL_VERSION.to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: (total_bytes * iterations) as u64,
-            items_processed: None,
-            messages_processed: Some((messages.len() * iterations) as u64),
-            artifact: capture_artifacts.then_some(ExecutionArtifact::IcuSummary(summary)),
-        })
-    }
-
-    fn run_node_adapter(
-        &self,
-        workspace: &Path,
-        scenario: &BenchmarkScenario,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let request = self.adapter_request(scenario, iterations, capture_artifacts);
-        let script = workspace.join("benchmark").join("node").join("adapter.cjs");
-        run_external_adapter(
-            "node",
-            &["--no-warnings", script.to_string_lossy().as_ref()],
-            workspace,
-            &request,
-        )
-    }
-
-    fn run_python_adapter(
-        &self,
-        workspace: &Path,
-        scenario: &BenchmarkScenario,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let request = self.adapter_request(scenario, iterations, capture_artifacts);
-        let script = workspace
-            .join("benchmark")
-            .join("python")
-            .join("adapter.py");
-        let python = preferred_python_program(workspace);
-        let args = vec![script.into_os_string()];
-        run_external_adapter(python.as_os_str(), &args, workspace, &request)
-    }
-
-    fn run_php_adapter(
-        &self,
-        workspace: &Path,
-        scenario: &BenchmarkScenario,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let request = self.adapter_request(scenario, iterations, capture_artifacts);
-        let script = workspace.join("benchmark").join("php").join("adapter.php");
-        let args = vec![script.into_os_string()];
-        run_external_adapter("php", &args, workspace, &request)
-    }
-
-    fn run_msgcat(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let input = self
-            .po_input_path
-            .as_ref()
-            .ok_or_else(|| "msgcat requires PO input path".to_owned())?;
-        let capture_path = self.tempdir.path().join("msgcat-output.po");
-        let start = Instant::now();
-        let mut last_stdout = Vec::new();
-        let mut bytes_processed = 0usize;
-        for _ in 0..iterations {
-            let output = Command::new("msgcat")
-                .arg("--no-wrap")
-                .arg(input)
-                .output()
-                .map_err(|error| format!("failed to launch msgcat: {error}"))?;
-            if !output.status.success() {
-                return Err(format!(
-                    "msgcat failed: {}",
-                    String::from_utf8_lossy(&output.stderr).trim()
-                ));
-            }
-            bytes_processed += output.stdout.len();
-            last_stdout = output.stdout;
-        }
-        let elapsed = start.elapsed();
-        let rendered = String::from_utf8(last_stdout)
-            .map_err(|error| format!("msgcat output was not valid UTF-8: {error}"))?;
-        if capture_artifacts {
-            fs::write(&capture_path, &rendered).map_err(|error| {
-                format!(
-                    "failed to persist msgcat output {}: {error}",
-                    capture_path.display()
-                )
-            })?;
-        }
-        let summary = PoSemanticSummary::from_po_file(
-            &parse_po(&rendered).map_err(|error| format!("msgcat output parse failed: {error}"))?,
-        );
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: read_command_version("msgcat", &["--version"])?
-                .lines()
-                .next()
-                .unwrap_or("msgcat")
-                .to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: bytes_processed as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: capture_artifacts.then_some(ExecutionArtifact::RenderedPoPath(capture_path)),
-        })
-    }
-
-    fn run_msgmerge(
-        &self,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> Result<ExecutionResult, String> {
-        let existing = self
-            .existing_po_path
-            .as_ref()
-            .ok_or_else(|| "msgmerge requires existing PO input".to_owned())?;
-        let pot = self
-            .pot_path
-            .as_ref()
-            .ok_or_else(|| "msgmerge requires a POT template path".to_owned())?;
-        let capture_path = self.tempdir.path().join("msgmerge-output.po");
-        let start = Instant::now();
-        let mut last_stdout = Vec::new();
-        let mut bytes_processed = 0usize;
-        for _ in 0..iterations {
-            let output = Command::new("msgmerge")
-                .arg("--no-wrap")
-                .arg("--quiet")
-                .arg("--no-fuzzy-matching")
-                .arg(existing)
-                .arg(pot)
-                .output()
-                .map_err(|error| format!("failed to launch msgmerge: {error}"))?;
-            if !output.status.success() {
-                return Err(format!(
-                    "msgmerge failed: {}",
-                    String::from_utf8_lossy(&output.stderr).trim()
-                ));
-            }
-            bytes_processed += output.stdout.len();
-            last_stdout = output.stdout;
-        }
-        let elapsed = start.elapsed();
-        let rendered = String::from_utf8(last_stdout)
-            .map_err(|error| format!("msgmerge output was not valid UTF-8: {error}"))?;
-        if capture_artifacts {
-            fs::write(&capture_path, &rendered).map_err(|error| {
-                format!(
-                    "failed to persist msgmerge output {}: {error}",
-                    capture_path.display()
-                )
-            })?;
-        }
-        let summary = PoSemanticSummary::from_po_file(
-            &parse_po(&rendered)
-                .map_err(|error| format!("msgmerge output parse failed: {error}"))?,
-        );
-        let digest = digest_summary(&summary)?;
-        Ok(ExecutionResult {
-            tool_version: read_command_version("msgmerge", &["--version"])?
-                .lines()
-                .next()
-                .unwrap_or("msgmerge")
-                .to_owned(),
-            reported_digest: digest,
-            elapsed_ns: elapsed.as_nanos(),
-            baseline_elapsed_ns: None,
-            bytes_processed: bytes_processed as u64,
-            items_processed: summary
-                .items
-                .len()
-                .checked_mul(iterations)
-                .map(|value| value as u64),
-            messages_processed: None,
-            artifact: capture_artifacts.then_some(ExecutionArtifact::RenderedPoPath(capture_path)),
-        })
-    }
-
-    fn adapter_request(
-        &self,
-        scenario: &BenchmarkScenario,
-        iterations: usize,
-        capture_artifacts: bool,
-    ) -> AdapterRequest {
-        AdapterRequest {
-            scenario_id: scenario.id.clone(),
-            implementation: scenario.implementation.clone(),
-            workload: scenario.workload.clone(),
-            operation: scenario.operation.clone(),
-            fixture: scenario.fixture.clone(),
-            iterations,
-            capture_artifacts,
-            po_input_path: self
-                .po_input_path
-                .as_ref()
-                .map(|path| path.to_string_lossy().into_owned()),
-            existing_po_path: self
-                .existing_po_path
-                .as_ref()
-                .map(|path| path.to_string_lossy().into_owned()),
-            pot_path: self
-                .pot_path
-                .as_ref()
-                .map(|path| path.to_string_lossy().into_owned()),
-            icu_messages_path: self
-                .icu_messages_path
-                .as_ref()
-                .map(|path| path.to_string_lossy().into_owned()),
-            po_output_path: capture_artifacts.then(|| {
-                self.tempdir
-                    .path()
-                    .join(format!("{}-output.po", scenario.implementation))
-                    .to_string_lossy()
-                    .into_owned()
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct OwnedMergeFixture {
-    existing_po: String,
-    api_messages: Vec<ExtractedMessage>,
-    template_pot: String,
-}
-
-impl OwnedMergeFixture {
-    fn from_fixture(fixture: &MergeFixture) -> Self {
-        Self {
-            existing_po: fixture.existing_po().to_owned(),
-            api_messages: fixture.api_extracted_messages().to_vec(),
-            // Same template the competitors parse, so the merge comparison is
-            // apples-to-apples: every tool reads existing.po and template.pot.
-            template_pot: build_merge_pot(fixture),
-        }
-    }
-}
-
-/// Builds the catalog-layer extraction input from a freshly parsed template,
-/// mirroring what a host extractor hands to `update_catalog`.
-fn extracted_messages_from_template(template: &PoFile) -> Vec<ExtractedMessage> {
-    template
-        .items
-        .iter()
-        .filter(|item| !item.obsolete)
-        .map(|item| {
-            let comments: Vec<String> = item.extracted_comments.iter().cloned().collect();
-            let origin: Vec<CatalogOrigin> = item
-                .references
-                .iter()
-                .map(|reference| parse_origin(reference))
-                .collect();
-            if let Some(msgid_plural) = item.msgid_plural.as_deref() {
-                ExtractedMessage::Plural(ExtractedPluralMessage {
-                    msgid: item.msgid.clone(),
-                    msgctxt: item.msgctxt.clone(),
-                    source: PluralSource {
-                        one: Some(item.msgid.clone()),
-                        other: msgid_plural.to_owned(),
-                    },
-                    comments,
-                    origin,
-                    placeholders: BTreeMap::default(),
-                })
-            } else {
-                ExtractedMessage::Singular(ExtractedSingularMessage {
-                    msgid: item.msgid.clone(),
-                    msgctxt: item.msgctxt.clone(),
-                    comments,
-                    origin,
-                    placeholders: BTreeMap::default(),
-                })
-            }
-        })
-        .collect()
-}
-
-fn build_merge_pot(fixture: &MergeFixture) -> String {
-    let mut out = String::new();
-    out.push_str("msgid \"\"\n");
-    out.push_str("msgstr \"\"\n");
-    out.push_str("\"Project-Id-Version: ferrocat benchmark template\\n\"\n");
-    out.push_str("\"Content-Type: text/plain; charset=UTF-8\\n\"\n\n");
-
-    for message in fixture.extracted_messages() {
-        for comment in &message.extracted_comments {
-            out.push_str("#. ");
-            out.push_str(comment);
-            out.push('\n');
-        }
-        if !message.references.is_empty() {
-            out.push_str("#: ");
-            let mut first = true;
-            for reference in &message.references {
-                if !first {
-                    out.push(' ');
-                }
-                first = false;
-                out.push_str(reference);
-            }
-            out.push('\n');
-        }
-        if !message.flags.is_empty() {
-            out.push_str("#, ");
-            let mut first = true;
-            for flag in &message.flags {
-                if !first {
-                    out.push_str(", ");
-                }
-                first = false;
-                out.push_str(flag);
-            }
-            out.push('\n');
-        }
-        if let Some(context) = &message.msgctxt {
-            push_po_keyword(&mut out, "msgctxt", context);
-        }
-        push_po_keyword(&mut out, "msgid", &message.msgid);
-        if let Some(plural) = &message.msgid_plural {
-            push_po_keyword(&mut out, "msgid_plural", plural);
-            out.push_str("msgstr[0] \"\"\n");
-            out.push_str("msgstr[1] \"\"\n");
-        } else {
-            out.push_str("msgstr \"\"\n");
-        }
-        out.push('\n');
-    }
-
-    out
-}
-
-fn push_po_keyword(out: &mut String, keyword: &str, value: &str) {
-    if !value.contains('\n') {
-        out.push_str(keyword);
-        out.push_str(" \"");
-        out.push_str(&escape_po_text(value));
-        out.push_str("\"\n");
-        return;
-    }
-
-    out.push_str(keyword);
-    out.push_str(" \"\"\n");
-    let mut lines = value.split('\n').peekable();
-    while let Some(line) = lines.next() {
-        out.push('"');
-        out.push_str(&escape_po_text(line));
-        if lines.peek().is_some() {
-            out.push_str("\\n");
-        }
-        out.push_str("\"\n");
-    }
-}
-
-fn escape_po_text(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    for ch in input.chars() {
-        match ch {
-            '"' => out.push_str("\\\""),
-            '\\' => out.push_str("\\\\"),
-            '\t' => out.push_str("\\t"),
-            '\r' => out.push_str("\\r"),
-            _ => out.push(ch),
-        }
-    }
-    out
-}
-
-fn load_fixture(name: &str) -> Result<Fixture, String> {
-    fixture_by_name(name).ok_or_else(|| format!("unknown benchmark fixture: {name}"))
-}
-
-fn load_icu_fixture(name: &str) -> Result<IcuFixture, String> {
-    icu_fixture_by_name(name).ok_or_else(|| format!("unknown ICU fixture: {name}"))
-}
-
-fn load_merge_fixture(name: &str) -> Result<MergeFixture, String> {
-    merge_fixture_by_name(name).ok_or_else(|| format!("unknown merge fixture: {name}"))
-}
-
-fn build_cli_baseline_po(locale: Option<&str>) -> String {
-    let (language, plural_forms) = fixture_locale_metadata(locale.unwrap_or("de"));
-    format!(
-        concat!(
-            "msgid \"\"\n",
-            "msgstr \"\"\n",
-            "\"Project-Id-Version: ferrocat benchmark baseline\\n\"\n",
-            "\"Language: {language}\\n\"\n",
-            "\"Content-Type: text/plain; charset=UTF-8\\n\"\n",
-            "\"Content-Transfer-Encoding: 8bit\\n\"\n",
-            "\"Plural-Forms: {plural_forms}\\n\"\n"
-        ),
-        language = language,
-        plural_forms = plural_forms
-    )
-}
-
-fn build_cli_baseline_pot() -> String {
-    concat!(
-        "msgid \"\"\n",
-        "msgstr \"\"\n",
-        "\"Project-Id-Version: ferrocat benchmark template\\n\"\n",
-        "\"Content-Type: text/plain; charset=UTF-8\\n\"\n",
-        "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-    )
-    .to_owned()
-}
-
-fn fixture_locale(name: &str) -> Option<String> {
-    if !name.starts_with("gettext-") {
-        return Some("de".to_owned());
-    }
-
-    let mut parts = name.split('-');
-    let _prefix = parts.next()?;
-    let _family = parts.next()?;
-    let locale = parts.next()?;
-    Some(locale.to_owned())
-}
-
-fn fixture_locale_metadata(locale: &str) -> (&'static str, &'static str) {
-    match locale {
-        "fr" => ("fr", "nplurals=2; plural=(n > 1);"),
-        "pl" => (
-            "pl",
-            "nplurals=3; plural=(n == 1 ? 0 : n % 10 >= 2 && n % 10 <= 4 && (n % 100 < 12 || n % 100 > 14) ? 1 : 2);",
-        ),
-        "ar" => (
-            "ar",
-            "nplurals=6; plural=(n==0 ? 0 : n==1 ? 1 : n==2 ? 2 : n%100>=3 && n%100<=10 ? 3 : n%100>=11 && n%100<=99 ? 4 : 5);",
-        ),
-        _ => ("de", "nplurals=2; plural=(n != 1);"),
-    }
-}
-
-fn fixture_catalog_mode(name: &str) -> CatalogMode {
-    if name.starts_with("gettext-") {
-        CatalogMode::GettextPo
-    } else {
-        CatalogMode::IcuPo
-    }
-}
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct PoSemanticSummary {
-    headers: Vec<PoHeaderSummary>,
-    items: Vec<PoItemSummary>,
-}
-
-impl PoSemanticSummary {
-    fn from_po_file(file: &PoFile) -> Self {
-        let headers = file
-            .headers
-            .iter()
-            .map(|header| PoHeaderSummary {
-                key: header.key.clone(),
-                value: header.value.clone(),
-            })
-            .collect::<Vec<_>>();
-        let items = file
-            .items
-            .iter()
-            .map(|item| PoItemSummary {
-                msgctxt: item.msgctxt.clone(),
-                msgid: item.msgid.clone(),
-                msgid_plural: item.msgid_plural.clone(),
-                msgstr: match &item.msgstr {
-                    MsgStr::None => Vec::new(),
-                    MsgStr::Singular(value) => vec![value.clone()],
-                    MsgStr::Plural(values) => values.clone(),
-                },
-                obsolete: item.obsolete,
-            })
-            .collect::<Vec<_>>();
-        Self { headers, items }.normalized()
-    }
-
-    fn from_borrowed_po_file(file: &BorrowedPoFile<'_>) -> Self {
-        let headers = file
-            .headers
-            .iter()
-            .map(|header| PoHeaderSummary {
-                key: header.key.as_ref().to_owned(),
-                value: header.value.as_ref().to_owned(),
-            })
-            .collect::<Vec<_>>();
-        let items = file
-            .items
-            .iter()
-            .map(|item| PoItemSummary {
-                msgctxt: item.msgctxt.as_ref().map(|value| value.as_ref().to_owned()),
-                msgid: item.msgid.as_ref().to_owned(),
-                msgid_plural: item
-                    .msgid_plural
-                    .as_ref()
-                    .map(|value| value.as_ref().to_owned()),
-                msgstr: match &item.msgstr {
-                    BorrowedMsgStr::None => Vec::new(),
-                    BorrowedMsgStr::Singular(value) => vec![value.as_ref().to_owned()],
-                    BorrowedMsgStr::Plural(values) => values
-                        .iter()
-                        .map(|value| value.as_ref().to_owned())
-                        .collect(),
-                },
-                obsolete: item.obsolete,
-            })
-            .collect::<Vec<_>>();
-        Self { headers, items }.normalized()
-    }
-
-    fn normalized(mut self) -> Self {
-        self.headers
-            .retain(|header| should_keep_benchmark_header(&header.key, &header.value));
-        self.headers.sort();
-        self.items.iter_mut().for_each(PoItemSummary::normalize);
-        self.items.sort();
-        self
-    }
-}
-
-fn should_keep_benchmark_header(key: &str, value: &str) -> bool {
-    !value.is_empty()
-        && !matches!(
-            key,
-            "MIME-Version" | "X-Generator" | "Content-Type" | "Content-Transfer-Encoding"
-        )
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct PoHeaderSummary {
-    key: String,
-    value: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct PoItemSummary {
-    msgctxt: Option<String>,
-    msgid: String,
-    msgid_plural: Option<String>,
-    msgstr: Vec<String>,
-    obsolete: bool,
-}
-
-impl PoItemSummary {
-    fn normalize(&mut self) {
-        if self.msgctxt.as_deref() == Some("") {
-            self.msgctxt = None;
-        }
-        if self.msgid_plural.as_deref() == Some("") {
-            self.msgid_plural = None;
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct CatalogSemanticSummary {
-    locale: Option<String>,
-    diagnostics: Vec<String>,
-    messages: Vec<CatalogMessageSummary>,
-}
-
-impl CatalogSemanticSummary {
-    fn from_parsed_catalog(parsed: ParsedCatalog) -> Result<Self, String> {
-        let normalized = parsed
-            .into_normalized_view()
-            .map_err(|error| format!("failed to normalize parsed catalog: {error}"))?;
-        let catalog = normalized.parsed_catalog();
-        let messages = normalized
-            .iter()
-            .map(|(_, message)| CatalogMessageSummary::from_message(message))
-            .collect::<Vec<_>>();
-        let diagnostics = catalog
-            .diagnostics
-            .iter()
-            .map(|diagnostic| {
-                format!(
-                    "{}:{}:{}",
-                    diagnostic_severity_label(diagnostic.severity),
-                    diagnostic.code,
-                    diagnostic.message
-                )
-            })
-            .collect::<Vec<_>>();
-
-        Ok(Self {
-            locale: catalog.locale.clone(),
-            diagnostics,
-            messages,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct CatalogMessageSummary {
-    msgctxt: Option<String>,
-    msgid: String,
-    translation: CatalogTranslationSummary,
-    comments: Vec<String>,
-    origins: Vec<CatalogOriginSummary>,
-    obsolete: bool,
-}
-
-impl CatalogMessageSummary {
-    fn from_message(message: &CatalogMessage) -> Self {
-        Self {
-            msgctxt: message.msgctxt.clone(),
-            msgid: message.msgid.clone(),
-            translation: CatalogTranslationSummary::from_translation(&message.translation),
-            comments: message.comments.clone(),
-            origins: message
-                .origin
-                .iter()
-                .map(CatalogOriginSummary::from_origin)
-                .collect(),
-            obsolete: message.obsolete.is_some(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-enum CatalogTranslationSummary {
-    Singular(String),
-    Plural {
-        source_one: Option<String>,
-        source_other: String,
-        variable: String,
-        forms: Vec<(String, String)>,
-    },
-}
-
-impl CatalogTranslationSummary {
-    fn from_translation(translation: &TranslationShape) -> Self {
-        match translation {
-            TranslationShape::Singular { value } => Self::Singular(value.clone()),
-            TranslationShape::Plural {
-                source,
-                translation,
-                variable,
-            } => Self::Plural {
-                source_one: source.one.clone(),
-                source_other: source.other.clone(),
-                variable: variable.clone(),
-                forms: translation
-                    .iter()
-                    .map(|(category, value)| (category.clone(), value.clone()))
-                    .collect(),
-            },
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct CatalogOriginSummary {
-    file: String,
-    scope: Option<String>,
-}
-
-impl CatalogOriginSummary {
-    fn from_origin(origin: &CatalogOrigin) -> Self {
-        Self {
-            file: origin.file.clone(),
-            scope: origin.scope.clone(),
-        }
-    }
-}
-
-fn diagnostic_severity_label(severity: ferrocat_po::DiagnosticSeverity) -> &'static str {
-    match severity {
-        ferrocat_po::DiagnosticSeverity::Info => "info",
-        ferrocat_po::DiagnosticSeverity::Warning => "warning",
-        ferrocat_po::DiagnosticSeverity::Error => "error",
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct IcuFixtureSummary {
-    messages: Vec<IcuMessageSummary>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct IcuMessageSummary {
-    variable_names: Vec<String>,
-    selector_kinds: Vec<String>,
-    selectors: Vec<String>,
-    plural_categories: Vec<String>,
-    tag_names: Vec<String>,
-    formatter_kinds: Vec<String>,
-    literal_segments: usize,
-    argument_count: usize,
-    pound_count: usize,
-    max_depth: usize,
-}
-
-impl IcuMessageSummary {
-    fn from_message(message: &IcuMessage) -> Self {
-        let mut collector = IcuCollector::default();
-        collector.visit_nodes(&message.nodes, 1);
-        Self {
-            variable_names: collector.variable_names.into_iter().collect(),
-            selector_kinds: collector.selector_kinds.into_iter().collect(),
-            selectors: collector.selectors.into_iter().collect(),
-            plural_categories: collector.plural_categories.into_iter().collect(),
-            tag_names: collector.tag_names.into_iter().collect(),
-            formatter_kinds: collector.formatter_kinds.into_iter().collect(),
-            literal_segments: collector.literal_segments,
-            argument_count: collector.argument_count,
-            pound_count: collector.pound_count,
-            max_depth: collector.max_depth,
-        }
-    }
-}
-
-#[derive(Default)]
-struct IcuCollector {
-    variable_names: BTreeSet<String>,
-    selector_kinds: BTreeSet<String>,
-    selectors: BTreeSet<String>,
-    plural_categories: BTreeSet<String>,
-    tag_names: BTreeSet<String>,
-    formatter_kinds: BTreeSet<String>,
-    literal_segments: usize,
-    argument_count: usize,
-    pound_count: usize,
-    max_depth: usize,
-}
-
-impl IcuCollector {
-    fn visit_nodes(&mut self, nodes: &[IcuNode], depth: usize) {
-        self.max_depth = self.max_depth.max(depth);
-        for node in nodes {
-            self.visit_node(node, depth);
-        }
-    }
-
-    fn visit_node(&mut self, node: &IcuNode, depth: usize) {
-        self.max_depth = self.max_depth.max(depth);
-        match node {
-            IcuNode::Literal(_) => self.literal_segments += 1,
-            IcuNode::Argument { name } => {
-                self.argument_count += 1;
-                self.variable_names.insert(name.clone());
-            }
-            IcuNode::Number { name, .. } => self.push_formatter("number", name),
-            IcuNode::Date { name, .. } => self.push_formatter("date", name),
-            IcuNode::Time { name, .. } => self.push_formatter("time", name),
-            IcuNode::List { name, .. } => self.push_formatter("list", name),
-            IcuNode::Duration { name, .. } => self.push_formatter("duration", name),
-            IcuNode::Ago { name, .. } => self.push_formatter("ago", name),
-            IcuNode::Name { name, .. } => self.push_formatter("name", name),
-            IcuNode::Select { name, options } => {
-                self.argument_count += 1;
-                self.variable_names.insert(name.clone());
-                self.selector_kinds.insert("select".to_owned());
-                self.visit_options(options, depth + 1, false);
-            }
-            IcuNode::Plural {
-                name,
-                kind,
-                options,
-                ..
-            } => {
-                self.argument_count += 1;
-                self.variable_names.insert(name.clone());
-                self.selector_kinds.insert(match kind {
-                    IcuPluralKind::Cardinal => "plural".to_owned(),
-                    IcuPluralKind::Ordinal => "selectordinal".to_owned(),
-                });
-                self.visit_options(options, depth + 1, true);
-            }
-            IcuNode::Pound => self.pound_count += 1,
-            IcuNode::Tag { name, children, .. } => {
-                self.tag_names.insert(name.clone());
-                self.visit_nodes(children, depth + 1);
-            }
-            _ => {}
-        }
-    }
-
-    fn visit_options(&mut self, options: &[IcuOption], depth: usize, plural: bool) {
-        for option in options {
-            self.selectors.insert(option.selector.clone());
-            if plural {
-                self.plural_categories.insert(option.selector.clone());
-            }
-            self.visit_nodes(&option.value, depth);
-        }
-    }
-
-    fn push_formatter(&mut self, kind: &str, name: &str) {
-        self.argument_count += 1;
-        self.variable_names.insert(name.to_owned());
-        self.formatter_kinds.insert(kind.to_owned());
-    }
-}
-
-fn digest_summary<T: Serialize>(value: &T) -> Result<String, String> {
-    let canonical = canonical_json_string(value)?;
-    let mut hasher = Sha256::new();
-    hasher.update(canonical.as_bytes());
-    let digest = hasher.finalize();
-    Ok(bytes_to_lower_hex(digest.as_ref()))
-}
-
-fn bytes_to_lower_hex(bytes: &[u8]) -> String {
-    const HEX: &[u8; 16] = b"0123456789abcdef";
-
-    let mut out = String::with_capacity(bytes.len() * 2);
-    for &byte in bytes {
-        out.push(HEX[usize::from(byte >> 4)] as char);
-        out.push(HEX[usize::from(byte & 0x0f)] as char);
-    }
-    out
-}
-
-fn canonical_json_string<T: Serialize>(value: &T) -> Result<String, String> {
-    let value = serde_json::to_value(value)
-        .map_err(|error| format!("failed to build canonical JSON value: {error}"))?;
-    let sorted = sort_json_value(value);
-    serde_json::to_string(&sorted)
-        .map_err(|error| format!("failed to render canonical JSON: {error}"))
-}
-
-fn sort_json_value(value: serde_json::Value) -> serde_json::Value {
-    match value {
-        serde_json::Value::Array(values) => {
-            serde_json::Value::Array(values.into_iter().map(sort_json_value).collect::<Vec<_>>())
-        }
-        serde_json::Value::Object(values) => {
-            let mut sorted = serde_json::Map::new();
-            let mut keys = values.into_iter().collect::<Vec<_>>();
-            keys.sort_by(|left, right| left.0.cmp(&right.0));
-            for (key, value) in keys {
-                sorted.insert(key, sort_json_value(value));
-            }
-            serde_json::Value::Object(sorted)
-        }
-        other => other,
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct CompareReport {
-    profile: String,
-    generated_at: String,
-    reference_host_policy: String,
-    environment: EnvironmentMetadata,
-    scenarios: Vec<ScenarioReport>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct ScenarioReport {
-    id: String,
-    comparison_group: String,
-    workload: String,
-    operation: String,
-    fixture: String,
-    implementation: String,
-    tool_version: String,
-    iterations_per_sample: usize,
-    warmup_runs: usize,
-    measured_runs: usize,
-    semantic_digest: String,
-    baseline_strategy: Option<String>,
-    baseline_fixture: Option<String>,
-    statistics: ScenarioStatistics,
-    samples: Vec<ScenarioSampleReport>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct ScenarioSampleReport {
-    elapsed_ns: u128,
-    baseline_elapsed_ns: Option<u128>,
-    adjusted_elapsed_ns: Option<u128>,
-    bytes_processed: u64,
-    items_processed: Option<u64>,
-    messages_processed: Option<u64>,
-    mib_per_sec: f64,
-    units_per_sec: f64,
-    adjusted_mib_per_sec: Option<f64>,
-    adjusted_units_per_sec: Option<f64>,
-}
-
-impl ScenarioSampleReport {
-    fn from_execution(sample: &ExecutionResult) -> Self {
-        let elapsed_seconds = nanos_to_seconds(sample.elapsed_ns);
-        let units = f64_from_u64(
-            sample
-                .items_processed
-                .or(sample.messages_processed)
-                .unwrap_or(0),
-        );
-        let adjusted_elapsed_ns = sample.adjusted_elapsed_ns();
-        let adjusted_seconds = adjusted_elapsed_ns.map(nanos_to_seconds);
-        Self {
-            elapsed_ns: sample.elapsed_ns,
-            baseline_elapsed_ns: sample.baseline_elapsed_ns,
-            adjusted_elapsed_ns,
-            bytes_processed: sample.bytes_processed,
-            items_processed: sample.items_processed,
-            messages_processed: sample.messages_processed,
-            mib_per_sec: throughput_mib(sample.bytes_processed, elapsed_seconds),
-            units_per_sec: throughput_units(units, elapsed_seconds),
-            adjusted_mib_per_sec: adjusted_seconds
-                .filter(|seconds| *seconds > 0.0)
-                .map(|seconds| throughput_mib(sample.bytes_processed, seconds)),
-            adjusted_units_per_sec: adjusted_seconds
-                .filter(|seconds| *seconds > 0.0)
-                .map(|seconds| throughput_units(units, seconds)),
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct ScenarioStatistics {
-    median_elapsed_ns: u128,
-    mean_elapsed_ns: f64,
-    min_elapsed_ns: u128,
-    max_elapsed_ns: u128,
-    stddev_elapsed_ns: f64,
-    median_absolute_deviation_ns: u128,
-    coefficient_of_variation: f64,
-    relative_span_percent: f64,
-    noisy: bool,
-    median_mib_per_sec: f64,
-    median_units_per_sec: f64,
-    median_baseline_elapsed_ns: Option<u128>,
-    median_adjusted_elapsed_ns: Option<u128>,
-    median_adjusted_mib_per_sec: Option<f64>,
-    median_adjusted_units_per_sec: Option<f64>,
-}
-
-impl ScenarioStatistics {
-    #[expect(
-        clippy::too_many_lines,
-        reason = "Benchmark statistics are computed together so the report contract stays easy to verify."
-    )]
-    fn from_samples(samples: &[ExecutionResult]) -> Self {
-        let mut elapsed = samples
-            .iter()
-            .map(|sample| sample.elapsed_ns)
-            .collect::<Vec<_>>();
-        elapsed.sort_unstable();
-        let median_elapsed_ns = elapsed[elapsed.len() / 2];
-        let min_elapsed_ns = *elapsed.first().unwrap_or(&0);
-        let max_elapsed_ns = *elapsed.last().unwrap_or(&0);
-        let mean = elapsed
-            .iter()
-            .map(|value| f64_from_u128(*value))
-            .sum::<f64>()
-            / f64_from_usize(elapsed.len());
-        let variance = elapsed
-            .iter()
-            .map(|value| {
-                let delta = f64_from_u128(*value) - mean;
-                delta * delta
-            })
-            .sum::<f64>()
-            / f64_from_usize(elapsed.len());
-        let median_absolute_deviation_ns = elapsed
-            .iter()
-            .map(|value| value.abs_diff(median_elapsed_ns))
-            .collect::<Vec<_>>();
-        let median_absolute_deviation_ns =
-            median_u128(median_absolute_deviation_ns).unwrap_or_default();
-        let stddev_elapsed_ns = variance.sqrt();
-        let coefficient_of_variation = if mean <= f64::EPSILON {
-            0.0
-        } else {
-            stddev_elapsed_ns / mean
-        };
-        let relative_span_percent = if median_elapsed_ns == 0 {
-            0.0
-        } else {
-            (f64_from_u128(max_elapsed_ns - min_elapsed_ns) / f64_from_u128(median_elapsed_ns))
-                * 100.0
-        };
-
-        let mut sample_reports = samples
-            .iter()
-            .map(|sample| {
-                let seconds = nanos_to_seconds(sample.elapsed_ns);
-                (
-                    throughput_mib(sample.bytes_processed, seconds),
-                    throughput_units(
-                        f64_from_u64(
-                            sample
-                                .items_processed
-                                .or(sample.messages_processed)
-                                .unwrap_or(0),
-                        ),
-                        seconds,
-                    ),
-                )
-            })
-            .collect::<Vec<_>>();
-        sample_reports.sort_by(|left, right| left.0.total_cmp(&right.0));
-        let median_mib_per_sec = sample_reports
-            .get(sample_reports.len() / 2)
-            .map_or(0.0, |entry| entry.0);
-        sample_reports.sort_by(|left, right| left.1.total_cmp(&right.1));
-        let median_units_per_sec = sample_reports
-            .get(sample_reports.len() / 2)
-            .map_or(0.0, |entry| entry.1);
-
-        let baseline_elapsed = samples
-            .iter()
-            .filter_map(|sample| sample.baseline_elapsed_ns)
-            .collect::<Vec<_>>();
-        let adjusted_elapsed = samples
-            .iter()
-            .filter_map(ExecutionResult::adjusted_elapsed_ns)
-            .collect::<Vec<_>>();
-        let adjusted_rates = samples
-            .iter()
-            .filter_map(|sample| {
-                let adjusted_ns = sample.adjusted_elapsed_ns()?;
-                if adjusted_ns == 0 {
-                    return None;
-                }
-                let seconds = nanos_to_seconds(adjusted_ns);
-                Some((
-                    throughput_mib(sample.bytes_processed, seconds),
-                    throughput_units(
-                        f64_from_u64(
-                            sample
-                                .items_processed
-                                .or(sample.messages_processed)
-                                .unwrap_or(0),
-                        ),
-                        seconds,
-                    ),
-                ))
-            })
-            .collect::<Vec<_>>();
-        let median_baseline_elapsed_ns = median_u128(baseline_elapsed);
-        let median_adjusted_elapsed_ns = median_u128(adjusted_elapsed);
-        let median_adjusted_mib_per_sec = median_f64(
-            adjusted_rates
-                .iter()
-                .map(|entry| entry.0)
-                .collect::<Vec<_>>(),
-        );
-        let median_adjusted_units_per_sec = median_f64(
-            adjusted_rates
-                .iter()
-                .map(|entry| entry.1)
-                .collect::<Vec<_>>(),
-        );
-
-        Self {
-            median_elapsed_ns,
-            mean_elapsed_ns: mean,
-            min_elapsed_ns,
-            max_elapsed_ns,
-            stddev_elapsed_ns,
-            median_absolute_deviation_ns,
-            coefficient_of_variation,
-            relative_span_percent,
-            noisy: coefficient_of_variation > NOISE_CV_WARNING_THRESHOLD
-                || relative_span_percent > NOISE_RELATIVE_SPAN_WARNING_THRESHOLD,
-            median_mib_per_sec,
-            median_units_per_sec,
-            median_baseline_elapsed_ns,
-            median_adjusted_elapsed_ns,
-            median_adjusted_mib_per_sec,
-            median_adjusted_units_per_sec,
-        }
-    }
-}
-
-fn median_u128(mut values: Vec<u128>) -> Option<u128> {
-    if values.is_empty() {
-        return None;
-    }
-    values.sort_unstable();
-    values.get(values.len() / 2).copied()
-}
-
-fn median_f64(mut values: Vec<f64>) -> Option<f64> {
-    if values.is_empty() {
-        return None;
-    }
-    values.sort_by(f64::total_cmp);
-    values.get(values.len() / 2).copied()
-}
-
-fn round_robin_schedule(run_counts: &[usize]) -> Vec<usize> {
-    let mut schedule = Vec::new();
-    let max_runs = run_counts.iter().copied().max().unwrap_or(0);
-    for round in 0..max_runs {
-        for (index, run_count) in run_counts.iter().copied().enumerate() {
-            if round < run_count {
-                schedule.push(index);
-            }
-        }
-    }
-    schedule
-}
-
-fn nanos_to_seconds(value: u128) -> f64 {
-    f64_from_u128(value) / 1_000_000_000.0
-}
-
-fn nanos_to_millis(value: u128) -> f64 {
-    f64_from_u128(value) / 1_000_000.0
-}
-
-fn throughput_mib(bytes: u64, seconds: f64) -> f64 {
-    if seconds <= 0.0 {
-        return f64::INFINITY;
-    }
-    f64_from_u64(bytes) / (1024.0 * 1024.0 * seconds)
-}
-
-fn throughput_units(units: f64, seconds: f64) -> f64 {
-    if seconds <= 0.0 {
-        return f64::INFINITY;
-    }
-    units / seconds
-}
-
-#[expect(
-    clippy::cast_precision_loss,
-    reason = "Benchmark summaries intentionally use approximate floating-point display values."
-)]
-const fn f64_from_u128(value: u128) -> f64 {
-    value as f64
-}
-
-#[expect(
-    clippy::cast_precision_loss,
-    reason = "Benchmark summaries intentionally use approximate floating-point display values."
-)]
-const fn f64_from_u64(value: u64) -> f64 {
-    value as f64
-}
-
-#[expect(
-    clippy::cast_precision_loss,
-    reason = "Benchmark summaries intentionally use approximate floating-point display values."
-)]
-const fn f64_from_usize(value: usize) -> f64 {
-    value as f64
-}
-
-#[derive(Debug, Clone)]
-struct ExecutionResult {
-    tool_version: String,
-    reported_digest: String,
-    elapsed_ns: u128,
-    baseline_elapsed_ns: Option<u128>,
-    bytes_processed: u64,
-    items_processed: Option<u64>,
-    messages_processed: Option<u64>,
-    artifact: Option<ExecutionArtifact>,
-}
-
-impl ExecutionResult {
-    fn adjusted_elapsed_ns(&self) -> Option<u128> {
-        self.baseline_elapsed_ns
-            .map(|baseline| self.elapsed_ns.saturating_sub(baseline))
-    }
-}
-
-#[derive(Debug, Clone)]
-enum ExecutionArtifact {
-    PoSummary(PoSemanticSummary),
-    CatalogSummary(CatalogSemanticSummary),
-    IcuSummary(IcuFixtureSummary),
-    RenderedPo(String),
-    RenderedPoPath(PathBuf),
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct EnvironmentMetadata {
-    git_sha: String,
-    system_label: String,
-    os: String,
-    cpu_model: String,
-    memory_bytes: u64,
-    rustc_version: String,
-    node_version: String,
-    python_version: String,
-    msgmerge_version: String,
-    msgcat_version: String,
-    node_adapter_version: String,
-    python_adapter_version: String,
-}
-
-#[derive(Debug)]
-struct BenchmarkEnvironment {
-    git_sha: String,
-    system_label: String,
-    os: String,
-    cpu_model: String,
-    memory_bytes: u64,
-    rustc_version: String,
-    node_version: String,
-    python_version: String,
-    msgmerge_version: String,
-    msgcat_version: String,
-    node_adapter_version: String,
-    python_adapter_version: String,
-    python_program: PathBuf,
-}
-
-impl BenchmarkEnvironment {
-    fn detect(
-        workspace: &Path,
-        path_override: Option<&OsStr>,
-        tool_requirement: ToolRequirement,
-    ) -> Result<Self, String> {
-        let mut errors = Vec::new();
-        let python_program = preferred_python_program(workspace);
-        let os = format!("{}-{}", env::consts::OS, env::consts::ARCH);
-        let cpu_model = detect_cpu_model(path_override);
-        let memory_bytes = detect_memory_bytes(path_override);
-
-        let rustc_version =
-            match read_command_version_with_path("rustc", &["--version"], path_override) {
-                Ok(version) => version,
-                Err(error) => {
-                    errors.push(error);
-                    String::new()
-                }
-            };
-        let mut external = ExternalToolVersions::not_required();
-        if tool_requirement == ToolRequirement::External {
-            external = detect_external_tool_versions(
-                workspace,
-                path_override,
-                &python_program,
-                &mut errors,
-            );
-        }
-
-        if !errors.is_empty() {
-            return Err(format!(
-                "benchmark environment verification failed:\n- {}",
-                errors.join("\n- ")
-            ));
-        }
-
-        Ok(Self {
-            git_sha: read_git_sha(workspace),
-            system_label: build_system_label(&cpu_model, memory_bytes),
-            os,
-            cpu_model,
-            memory_bytes,
-            rustc_version,
-            node_version: external.node_version,
-            python_version: external.python_version,
-            msgmerge_version: external.msgmerge_version,
-            msgcat_version: external.msgcat_version,
-            node_adapter_version: external.node_adapter_version,
-            python_adapter_version: external.python_adapter_version,
-            python_program,
-        })
-    }
-
-    fn metadata(&self) -> EnvironmentMetadata {
-        EnvironmentMetadata {
-            git_sha: self.git_sha.clone(),
-            system_label: self.system_label.clone(),
-            os: self.os.clone(),
-            cpu_model: self.cpu_model.clone(),
-            memory_bytes: self.memory_bytes,
-            rustc_version: self.rustc_version.clone(),
-            node_version: self.node_version.clone(),
-            python_version: self.python_version.clone(),
-            msgmerge_version: self.msgmerge_version.clone(),
-            msgcat_version: self.msgcat_version.clone(),
-            node_adapter_version: self.node_adapter_version.clone(),
-            python_adapter_version: self.python_adapter_version.clone(),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct ExternalToolVersions {
-    node_version: String,
-    python_version: String,
-    msgmerge_version: String,
-    msgcat_version: String,
-    node_adapter_version: String,
-    python_adapter_version: String,
-}
-
-impl ExternalToolVersions {
-    fn not_required() -> Self {
-        Self {
-            node_version: "not-required".to_owned(),
-            python_version: "not-required".to_owned(),
-            msgmerge_version: "not-required".to_owned(),
-            msgcat_version: "not-required".to_owned(),
-            node_adapter_version: "not-required".to_owned(),
-            python_adapter_version: "not-required".to_owned(),
-        }
-    }
-}
-
-fn detect_external_tool_versions(
-    workspace: &Path,
-    path_override: Option<&OsStr>,
-    python_program: &Path,
-    errors: &mut Vec<String>,
-) -> ExternalToolVersions {
-    let node_version = match read_command_version_with_path("node", &["--version"], path_override) {
-        Ok(version) => version,
-        Err(error) => {
-            errors.push(error);
-            String::new()
-        }
-    };
-    let python_version = match read_command_version_for_program(
-        python_program,
-        &["--version"],
-        workspace,
-        path_override,
-    ) {
-        Ok(version) => version,
-        Err(error) => {
-            errors.push(error);
-            String::new()
-        }
-    };
-    let msgmerge_version =
-        match read_command_version_with_path("msgmerge", &["--version"], path_override) {
-            Ok(version) => version,
-            Err(error) => {
-                errors.push(error);
-                String::new()
-            }
-        };
-    let msgcat_version =
-        match read_command_version_with_path("msgcat", &["--version"], path_override) {
-            Ok(version) => version,
-            Err(error) => {
-                errors.push(error);
-                String::new()
-            }
-        };
-    let node_adapter_version = match run_command_capture_with_path(
-        "node",
-        &[
-            OsString::from("--no-warnings"),
-            workspace
-                .join("benchmark")
-                .join("node")
-                .join("adapter.cjs")
-                .into_os_string(),
-            OsString::from("--check"),
-        ],
-        workspace,
-        path_override,
-    ) {
-        Ok(output) => output.stdout.trim().to_owned(),
-        Err(error) => {
-            errors.push(error);
-            String::new()
-        }
-    };
-    let python_adapter_version = match run_command_capture_with_path(
-        python_program.as_os_str(),
-        &[
-            workspace
-                .join("benchmark")
-                .join("python")
-                .join("adapter.py")
-                .into_os_string(),
-            OsString::from("--check"),
-        ],
-        workspace,
-        path_override,
-    ) {
-        Ok(output) => output.stdout.trim().to_owned(),
-        Err(error) => {
-            errors.push(error);
-            String::new()
-        }
-    };
-
-    ExternalToolVersions {
-        node_version,
-        python_version,
-        msgmerge_version,
-        msgcat_version,
-        node_adapter_version,
-        python_adapter_version,
-    }
-}
-
-fn read_git_sha(workspace: &Path) -> String {
-    match run_command_capture("git", &["rev-parse", "HEAD"], workspace) {
-        Ok(output) => output.stdout.trim().to_owned(),
-        Err(_) => "unknown".to_owned(),
-    }
-}
-
-fn detect_cpu_model(path_override: Option<&OsStr>) -> String {
-    let workspace = workspace_root().unwrap_or_else(|_| PathBuf::from("."));
-    if env::consts::OS == "macos" {
-        if let Some(value) = read_macos_sysctl_string("machdep.cpu.brand_string") {
-            return value;
-        }
-        if let Ok(output) =
-            run_command_capture_with_path("uname", &["-m"], &workspace, path_override)
-        {
-            let value = output.stdout.trim();
-            if !value.is_empty() {
-                return value.to_owned();
-            }
-        }
-    }
-    if let Ok(output) =
-        run_command_capture_with_path("lscpu", &[] as &[&str], &workspace, path_override)
-    {
-        for line in output.stdout.lines() {
-            if let Some(value) = line.strip_prefix("Model name:") {
-                let trimmed = value.trim();
-                if !trimmed.is_empty() {
-                    return trimmed.to_owned();
-                }
-            }
-        }
-    }
-    if let Ok(output) = run_command_capture_with_path("uname", &["-m"], &workspace, path_override) {
-        let value = output.stdout.trim();
-        if !value.is_empty() {
-            return value.to_owned();
-        }
-    }
-    "unknown-cpu".to_owned()
-}
-
-fn detect_memory_bytes(path_override: Option<&OsStr>) -> u64 {
-    let workspace = workspace_root().unwrap_or_else(|_| PathBuf::from("."));
-    if env::consts::OS == "macos"
-        && let Some(bytes) = read_macos_sysctl_u64("hw.memsize")
-    {
-        return bytes;
-    }
-    if env::consts::OS == "linux"
-        && let Ok(meminfo) = fs::read_to_string("/proc/meminfo")
-    {
-        for line in meminfo.lines() {
-            if let Some(kb) = line
-                .strip_prefix("MemTotal:")
-                .and_then(|value| value.split_whitespace().next())
-                .and_then(|raw| raw.parse::<u64>().ok())
-            {
-                return kb.saturating_mul(1024);
-            }
-        }
-    }
-    if env::consts::OS == "windows"
-        && let Ok(output) = run_command_capture_with_path(
-            "powershell",
-            &[
-                "-NoProfile",
-                "-Command",
-                "(Get-CimInstance Win32_ComputerSystem).TotalPhysicalMemory",
-            ],
-            &workspace,
-            path_override,
-        )
-        && let Ok(bytes) = output.stdout.trim().parse::<u64>()
-    {
-        return bytes;
-    }
-    0
-}
-
-fn build_system_label(cpu_model: &str, memory_bytes: u64) -> String {
-    build_system_label_with_os(cpu_model, memory_bytes, &human_os_label())
-}
-
-fn build_system_label_with_os(cpu_model: &str, memory_bytes: u64, os: &str) -> String {
-    let cpu = if cpu_model.trim().is_empty() {
-        "Unknown CPU"
-    } else {
-        cpu_model.trim()
-    };
-    if memory_bytes == 0 {
-        return format!("{cpu} ({os})");
-    }
-    format!("{cpu} ({}, {os})", format_memory_label(memory_bytes))
-}
-
-fn human_os_label() -> String {
-    let os = match env::consts::OS {
-        "macos" => "macOS",
-        "windows" => "Windows",
-        "linux" => "Linux",
-        other => other,
-    };
-    let arch = match env::consts::ARCH {
-        "aarch64" => "arm64",
-        other => other,
-    };
-    format!("{os} {arch}")
-}
-
-fn format_memory_label(memory_bytes: u64) -> String {
-    if memory_bytes == 0 {
-        return "unknown RAM".to_owned();
-    }
-
-    let gib = f64_from_u64(memory_bytes) / 1024_f64.powi(3);
-    let rounded = gib.round();
-    if (gib - rounded).abs() < 0.05 {
-        format!("{rounded:.0} GB RAM")
-    } else {
-        format!("{gib:.1} GB RAM")
-    }
-}
-
-#[cfg(target_os = "macos")]
-fn read_macos_sysctl_string(name: &str) -> Option<String> {
-    let name = CString::new(name).ok()?;
-    let mut len = 0usize;
-    // SAFETY: `name` is a valid NUL-terminated C string, the first call asks macOS
-    // for the required buffer size, and the second call writes into an owned buffer
-    // of that exact size.
-    unsafe {
-        if libc::sysctlbyname(
-            name.as_ptr(),
-            std::ptr::null_mut(),
-            &raw mut len,
-            std::ptr::null_mut(),
-            0,
-        ) != 0
-            || len == 0
-        {
-            return None;
-        }
-
-        let mut buffer = vec![0u8; len];
-        if libc::sysctlbyname(
-            name.as_ptr(),
-            buffer.as_mut_ptr().cast(),
-            &raw mut len,
-            std::ptr::null_mut(),
-            0,
-        ) != 0
-        {
-            return None;
-        }
-
-        if len > 0 {
-            buffer.truncate(len.saturating_sub(1));
-        }
-        String::from_utf8(buffer)
-            .ok()
-            .map(|value| value.trim().to_owned())
-            .filter(|value| !value.is_empty())
-    }
-}
-
-#[cfg(not(target_os = "macos"))]
-fn read_macos_sysctl_string(_name: &str) -> Option<String> {
-    None
-}
-
-#[cfg(target_os = "macos")]
-fn read_macos_sysctl_u64(name: &str) -> Option<u64> {
-    let name = CString::new(name).ok()?;
-    let mut value = 0u64;
-    let mut len = std::mem::size_of::<u64>();
-    // SAFETY: `name` is a valid NUL-terminated C string and `value` points to a
-    // properly sized writable `u64` buffer for `sysctlbyname`.
-    unsafe {
-        if libc::sysctlbyname(
-            name.as_ptr(),
-            (&raw mut value).cast(),
-            &raw mut len,
-            std::ptr::null_mut(),
-            0,
-        ) != 0
-            || len != std::mem::size_of::<u64>()
-        {
-            return None;
-        }
-    }
-    Some(value)
-}
-
-#[cfg(not(target_os = "macos"))]
-fn read_macos_sysctl_u64(_name: &str) -> Option<u64> {
-    None
-}
-
-fn read_command_version(program: &str, args: &[&str]) -> Result<String, String> {
-    read_command_version_with_path(program, args, None)
-}
-
-fn read_command_version_for_program(
-    program: &Path,
-    args: &[&str],
-    cwd: &Path,
-    path_override: Option<&OsStr>,
-) -> Result<String, String> {
-    let output = run_command_capture_with_path(program.as_os_str(), args, cwd, path_override)?;
-    let version = output.stdout.trim();
-    if version.is_empty() {
-        return Err(format!("{} produced no version output", program.display()));
-    }
-    Ok(version.to_owned())
-}
-
-fn read_command_version_with_path(
-    program: &str,
-    args: &[&str],
-    path_override: Option<&OsStr>,
-) -> Result<String, String> {
-    let workspace = workspace_root()?;
-    let output = run_command_capture_with_path(program, args, &workspace, path_override)?;
-    let version = output.stdout.trim();
-    if version.is_empty() {
-        return Err(format!("{program} produced no version output"));
-    }
-    Ok(version.to_owned())
-}
-
-#[derive(Debug)]
-struct CommandCapture {
-    stdout: String,
-}
-
-fn run_command_capture(program: &str, args: &[&str], cwd: &Path) -> Result<CommandCapture, String> {
-    run_command_capture_with_path(program, args, cwd, None)
-}
-
-fn run_command_capture_with_path(
-    program: impl AsRef<OsStr>,
-    args: &[impl AsRef<OsStr>],
-    cwd: &Path,
-    path_override: Option<&OsStr>,
-) -> Result<CommandCapture, String> {
-    let program = program.as_ref();
-    let program_label = program.to_string_lossy().into_owned();
-    let mut command = Command::new(program);
-    command.current_dir(cwd);
-    if let Some(path_override) = path_override {
-        command.env("PATH", path_override);
-    }
-    for arg in args {
-        command.arg(arg);
-    }
-
-    let output = command
-        .output()
-        .map_err(|error| format!("failed to launch {program_label}: {error}"))?;
-    if !output.status.success() {
-        return Err(format!(
-            "{program_label} exited with {}: {}",
-            output.status,
-            String::from_utf8_lossy(&output.stderr).trim()
-        ));
-    }
-    Ok(CommandCapture {
-        stdout: String::from_utf8_lossy(&output.stdout).into_owned(),
-    })
-}
-
-fn preferred_python_program(workspace: &Path) -> PathBuf {
-    let candidates = [
-        workspace
-            .join("benchmark")
-            .join("python")
-            .join(".venv")
-            .join("bin")
-            .join("python3"),
-        workspace
-            .join("benchmark")
-            .join("python")
-            .join(".venv")
-            .join("bin")
-            .join("python"),
-        workspace
-            .join("benchmark")
-            .join("python")
-            .join(".venv")
-            .join("Scripts")
-            .join("python.exe"),
-    ];
-
-    for candidate in candidates {
-        if candidate.is_file() {
-            return candidate;
-        }
-    }
-
-    PathBuf::from("python3")
-}
-
-#[derive(Debug, Serialize)]
-struct AdapterRequest {
-    scenario_id: String,
-    implementation: String,
-    workload: String,
-    operation: String,
-    fixture: String,
-    iterations: usize,
-    capture_artifacts: bool,
-    po_input_path: Option<String>,
-    existing_po_path: Option<String>,
-    pot_path: Option<String>,
-    icu_messages_path: Option<String>,
-    po_output_path: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct AdapterResponse {
-    implementation: String,
-    workload: String,
-    fixture: String,
-    success: bool,
-    semantic_digest: String,
-    elapsed_ns: u128,
-    bytes_processed: u64,
-    items_processed: Option<u64>,
-    messages_processed: Option<u64>,
-    tool_version: String,
-    po_summary: Option<PoSemanticSummary>,
-    icu_summary: Option<IcuFixtureSummary>,
-    po_output_path: Option<String>,
-}
-
-fn run_external_adapter(
-    program: impl AsRef<OsStr>,
-    args: &[impl AsRef<OsStr>],
-    workspace: &Path,
-    request: &AdapterRequest,
-) -> Result<ExecutionResult, String> {
-    let input = serde_json::to_vec(request)
-        .map_err(|error| format!("failed to serialize adapter request: {error}"))?;
-    let program = program.as_ref();
-    let program_label = program.to_string_lossy().into_owned();
-    let mut command = Command::new(program);
-    command.current_dir(workspace);
-    for arg in args {
-        command.arg(arg);
-    }
-    command.stdin(std::process::Stdio::piped());
-    command.stdout(std::process::Stdio::piped());
-    command.stderr(std::process::Stdio::piped());
-    let mut child = command
-        .spawn()
-        .map_err(|error| format!("failed to launch {program_label}: {error}"))?;
-    let Some(mut stdin) = child.stdin.take() else {
-        return Err(format!("failed to open stdin for {program_label}"));
-    };
-    stdin
-        .write_all(&input)
-        .map_err(|error| format!("failed to write adapter request: {error}"))?;
-    drop(stdin);
-
-    let output = child
-        .wait_with_output()
-        .map_err(|error| format!("failed to wait for {program_label}: {error}"))?;
-    if !output.status.success() {
-        return Err(format!(
-            "{program_label} adapter failed: {}",
-            String::from_utf8_lossy(&output.stderr).trim()
-        ));
-    }
-    let response: AdapterResponse = serde_json::from_slice(&output.stdout)
-        .map_err(|error| format!("failed to parse adapter response: {error}"))?;
-    if !response.success {
-        return Err(format!(
-            "{program_label} adapter reported unsuccessful execution for scenario {}",
-            request.scenario_id
-        ));
-    }
-    if response.implementation != request.implementation
-        || response.workload != request.workload
-        || response.fixture != request.fixture
-    {
-        return Err(format!(
-            "{program_label} adapter response metadata mismatch for scenario {}",
-            request.scenario_id
-        ));
-    }
-
-    let artifact = if let Some(summary) = response.po_summary {
-        Some(ExecutionArtifact::PoSummary(summary.normalized()))
-    } else if let Some(summary) = response.icu_summary {
-        Some(ExecutionArtifact::IcuSummary(summary))
-    } else {
-        response
-            .po_output_path
-            .map(|path| ExecutionArtifact::RenderedPoPath(PathBuf::from(path)))
-    };
-
-    Ok(ExecutionResult {
-        tool_version: response.tool_version,
-        reported_digest: response.semantic_digest,
-        elapsed_ns: response.elapsed_ns,
-        baseline_elapsed_ns: None,
-        bytes_processed: response.bytes_processed,
-        items_processed: response.items_processed,
-        messages_processed: response.messages_processed,
-        artifact,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn first_po_summary_difference(left: &PoSemanticSummary, right: &PoSemanticSummary) -> String {
-        if left.headers != right.headers {
-            return format!(
-                "headers differ\nleft={}\nright={}",
-                canonical_json_string(&left.headers)
-                    .unwrap_or_else(|_| "<left-json-error>".to_owned()),
-                canonical_json_string(&right.headers)
-                    .unwrap_or_else(|_| "<right-json-error>".to_owned())
-            );
-        }
-
-        if left.items.len() != right.items.len() {
-            return format!(
-                "item count differs: left={} right={}",
-                left.items.len(),
-                right.items.len()
-            );
-        }
-
-        for (index, (left_item, right_item)) in left.items.iter().zip(&right.items).enumerate() {
-            if left_item != right_item {
-                return format!(
-                    "item {} differs\nleft={}\nright={}",
-                    index,
-                    canonical_json_string(left_item)
-                        .unwrap_or_else(|_| "<left-json-error>".to_owned()),
-                    canonical_json_string(right_item)
-                        .unwrap_or_else(|_| "<right-json-error>".to_owned())
-                );
-            }
-        }
-
-        "no summary difference found".to_owned()
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_polib_gettext_ui_de_compatibility() {
-        let workspace = workspace_root().expect("workspace");
-        let scenarios = vec![
-            BenchmarkScenario {
-                id: "po-parse/gettext-ui-de-1000/ferrocat-owned".to_owned(),
-                comparison_group: "po-parse/gettext-ui-de-1000".to_owned(),
-                workload: "po-parse".to_owned(),
-                operation: "parse".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "ferrocat-parse".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-            BenchmarkScenario {
-                id: "po-parse/gettext-ui-de-1000/polib".to_owned(),
-                comparison_group: "po-parse/gettext-ui-de-1000".to_owned(),
-                workload: "po-parse".to_owned(),
-                operation: "parse".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "polib".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-        ];
-
-        let prepared = PreparedScenario::prepare(&workspace, &scenarios).expect("prepared");
-        let internal = execute_scenario(&workspace, &prepared, &scenarios[0], 1, true)
-            .expect("internal parse");
-        let polib =
-            execute_scenario(&workspace, &prepared, &scenarios[1], 1, true).expect("polib parse");
-
-        let ExecutionArtifact::PoSummary(internal_summary) =
-            internal.artifact.expect("internal artifact")
-        else {
-            panic!("internal scenario did not return a po summary");
-        };
-        let ExecutionArtifact::PoSummary(polib_summary) = polib.artifact.expect("polib artifact")
-        else {
-            panic!("polib scenario did not return a po summary");
-        };
-
-        assert_eq!(
-            internal_summary,
-            polib_summary,
-            "{}",
-            first_po_summary_difference(&internal_summary, &polib_summary)
-        );
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_pofile_gettext_ui_de_compatibility() {
-        let workspace = workspace_root().expect("workspace");
-        let scenarios = vec![
-            BenchmarkScenario {
-                id: "po-parse/gettext-ui-de-1000/ferrocat-owned".to_owned(),
-                comparison_group: "po-parse/gettext-ui-de-1000".to_owned(),
-                workload: "po-parse".to_owned(),
-                operation: "parse".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "ferrocat-parse".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-            BenchmarkScenario {
-                id: "po-parse/gettext-ui-de-1000/pofile".to_owned(),
-                comparison_group: "po-parse/gettext-ui-de-1000".to_owned(),
-                workload: "po-parse".to_owned(),
-                operation: "parse".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "pofile".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-        ];
-
-        let prepared = PreparedScenario::prepare(&workspace, &scenarios).expect("prepared");
-        let internal = execute_scenario(&workspace, &prepared, &scenarios[0], 1, true)
-            .expect("internal parse");
-        let pofile =
-            execute_scenario(&workspace, &prepared, &scenarios[1], 1, true).expect("pofile parse");
-
-        let ExecutionArtifact::PoSummary(internal_summary) =
-            internal.artifact.expect("internal artifact")
-        else {
-            panic!("internal scenario did not return a po summary");
-        };
-        let ExecutionArtifact::PoSummary(pofile_summary) =
-            pofile.artifact.expect("pofile artifact")
-        else {
-            panic!("pofile scenario did not return a po summary");
-        };
-
-        assert_eq!(
-            internal_summary,
-            pofile_summary,
-            "{}",
-            first_po_summary_difference(&internal_summary, &pofile_summary)
-        );
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_msgmerge_gettext_ui_de_merge_compatibility() {
-        let workspace = workspace_root().expect("workspace");
-        let scenarios = vec![
-            BenchmarkScenario {
-                id: "po-merge/gettext-ui-de-1000/ferrocat".to_owned(),
-                comparison_group: "po-merge/gettext-ui-de-1000".to_owned(),
-                workload: "po-merge-update".to_owned(),
-                operation: "merge".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "ferrocat-merge".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-            BenchmarkScenario {
-                id: "po-merge/gettext-ui-de-1000/msgmerge".to_owned(),
-                comparison_group: "po-merge/gettext-ui-de-1000".to_owned(),
-                workload: "po-merge-update".to_owned(),
-                operation: "merge".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "msgmerge".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-        ];
-
-        let prepared = PreparedScenario::prepare(&workspace, &scenarios).expect("prepared");
-        let internal = execute_scenario(&workspace, &prepared, &scenarios[0], 1, true)
-            .expect("internal merge");
-        let external =
-            execute_scenario(&workspace, &prepared, &scenarios[1], 1, true).expect("msgmerge");
-
-        let internal_rendered = match internal.artifact.expect("internal artifact") {
-            ExecutionArtifact::RenderedPo(content) => content,
-            ExecutionArtifact::RenderedPoPath(path) => {
-                std::fs::read_to_string(path).expect("read internal rendered output")
-            }
-            other => panic!("unexpected internal artifact: {other:?}"),
-        };
-        let external_rendered = match external.artifact.expect("external artifact") {
-            ExecutionArtifact::RenderedPo(content) => content,
-            ExecutionArtifact::RenderedPoPath(path) => {
-                std::fs::read_to_string(path).expect("read external rendered output")
-            }
-            other => panic!("unexpected external artifact: {other:?}"),
-        };
-
-        let internal_summary =
-            PoSemanticSummary::from_po_file(&parse_po(&internal_rendered).expect("parse internal"));
-        let external_summary =
-            PoSemanticSummary::from_po_file(&parse_po(&external_rendered).expect("parse external"));
-
-        assert_eq!(
-            internal_summary,
-            external_summary,
-            "{}",
-            first_po_summary_difference(&internal_summary, &external_summary)
-        );
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_msgmerge_gettext_ui_de_update_compatibility() {
-        let workspace = workspace_root().expect("workspace");
-        let scenarios = vec![
-            BenchmarkScenario {
-                id: "po-update/gettext-ui-de-1000/ferrocat".to_owned(),
-                comparison_group: "po-update/gettext-ui-de-1000".to_owned(),
-                workload: "po-merge-update".to_owned(),
-                operation: "update-catalog".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "ferrocat-update-catalog".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-            BenchmarkScenario {
-                id: "po-update/gettext-ui-de-1000/ferrocat-file".to_owned(),
-                comparison_group: "po-update/gettext-ui-de-1000".to_owned(),
-                workload: "po-merge-update".to_owned(),
-                operation: "update-catalog".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "ferrocat-update-catalog-file".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-            BenchmarkScenario {
-                id: "po-update/gettext-ui-de-1000/msgmerge".to_owned(),
-                comparison_group: "po-update/gettext-ui-de-1000".to_owned(),
-                workload: "po-merge-update".to_owned(),
-                operation: "update-catalog".to_owned(),
-                fixture: "gettext-ui-de-1000".to_owned(),
-                implementation: "msgmerge".to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-        ];
-
-        let prepared = PreparedScenario::prepare(&workspace, &scenarios).expect("prepared");
-        let internal = execute_scenario(&workspace, &prepared, &scenarios[0], 1, true)
-            .expect("internal update");
-        let internal_file = execute_scenario(&workspace, &prepared, &scenarios[1], 1, true)
-            .expect("internal file-based update");
-        let external =
-            execute_scenario(&workspace, &prepared, &scenarios[2], 1, true).expect("msgmerge");
-
-        assert_eq!(
-            internal.reported_digest, internal_file.reported_digest,
-            "file-based catalog update must match the pre-structured variant"
-        );
-
-        let internal_rendered = match internal.artifact.expect("internal artifact") {
-            ExecutionArtifact::RenderedPo(content) => content,
-            ExecutionArtifact::RenderedPoPath(path) => {
-                std::fs::read_to_string(path).expect("read internal rendered output")
-            }
-            other => panic!("unexpected internal artifact: {other:?}"),
-        };
-        let external_rendered = match external.artifact.expect("external artifact") {
-            ExecutionArtifact::RenderedPo(content) => content,
-            ExecutionArtifact::RenderedPoPath(path) => {
-                std::fs::read_to_string(path).expect("read external rendered output")
-            }
-            other => panic!("unexpected external artifact: {other:?}"),
-        };
-
-        let internal_summary =
-            PoSemanticSummary::from_po_file(&parse_po(&internal_rendered).expect("parse internal"));
-        let external_summary =
-            PoSemanticSummary::from_po_file(&parse_po(&external_rendered).expect("parse external"));
-
-        assert_eq!(
-            internal_summary,
-            external_summary,
-            "{}",
-            first_po_summary_difference(&internal_summary, &external_summary)
-        );
-    }
-
-    fn debug_external_workflow_compatibility(operation: &str, implementation: &str) {
-        let workspace = workspace_root().expect("workspace");
-        let internal_impl = if operation == "merge" {
-            "ferrocat-merge"
-        } else {
-            "ferrocat-update-catalog"
-        };
-        let fixture = "gettext-ui-de-1000";
-        let group_prefix = if operation == "merge" {
-            "po-merge"
-        } else {
-            "po-update"
-        };
-        let scenarios = vec![
-            BenchmarkScenario {
-                id: format!("{group_prefix}/{fixture}/internal"),
-                comparison_group: format!("{group_prefix}/{fixture}"),
-                workload: "po-merge-update".to_owned(),
-                operation: operation.to_owned(),
-                fixture: fixture.to_owned(),
-                implementation: internal_impl.to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-            BenchmarkScenario {
-                id: format!("{group_prefix}/{fixture}/{implementation}"),
-                comparison_group: format!("{group_prefix}/{fixture}"),
-                workload: "po-merge-update".to_owned(),
-                operation: operation.to_owned(),
-                fixture: fixture.to_owned(),
-                implementation: implementation.to_owned(),
-                warmup_runs: 0,
-                measured_runs: 1,
-                minimum_sample_millis: Some(1),
-            },
-        ];
-
-        let prepared = PreparedScenario::prepare(&workspace, &scenarios).expect("prepared");
-        let internal = execute_scenario(&workspace, &prepared, &scenarios[0], 1, true)
-            .expect("internal workflow");
-        let external = execute_scenario(&workspace, &prepared, &scenarios[1], 1, true)
-            .expect("external workflow");
-
-        let internal_rendered = match internal.artifact.expect("internal artifact") {
-            ExecutionArtifact::RenderedPo(content) => content,
-            ExecutionArtifact::RenderedPoPath(path) => {
-                std::fs::read_to_string(path).expect("read internal rendered output")
-            }
-            other => panic!("unexpected internal artifact: {other:?}"),
-        };
-        let external_rendered = match external.artifact.expect("external artifact") {
-            ExecutionArtifact::RenderedPo(content) => content,
-            ExecutionArtifact::RenderedPoPath(path) => {
-                std::fs::read_to_string(path).expect("read external rendered output")
-            }
-            other => panic!("unexpected external artifact: {other:?}"),
-        };
-
-        let internal_summary =
-            PoSemanticSummary::from_po_file(&parse_po(&internal_rendered).expect("parse internal"));
-        let external_summary =
-            PoSemanticSummary::from_po_file(&parse_po(&external_rendered).expect("parse external"));
-
-        assert_eq!(
-            external.reported_digest,
-            digest_summary(&external_summary).expect("external digest"),
-            "{}",
-            first_po_summary_difference(
-                &PoSemanticSummary::from_po_file(
-                    &parse_po(&external_rendered).expect("reparse external"),
-                ),
-                &external_summary
-            )
-        );
-
-        assert_eq!(
-            internal_summary,
-            external_summary,
-            "{}",
-            first_po_summary_difference(&internal_summary, &external_summary)
-        );
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_pofile_gettext_ui_de_merge_workflow_compatibility() {
-        debug_external_workflow_compatibility("merge", "pofile");
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_pofile_ts_gettext_ui_de_merge_workflow_compatibility() {
-        debug_external_workflow_compatibility("merge", "pofile-ts");
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_polib_gettext_ui_de_merge_workflow_compatibility() {
-        debug_external_workflow_compatibility("merge", "polib");
-    }
-
-    #[test]
-    #[ignore = "manual compatibility probe for external adapters"]
-    fn debug_babel_gettext_ui_de_update_workflow_compatibility() {
-        debug_external_workflow_compatibility("update-catalog", "babel");
-    }
-
-    #[test]
-    fn canonical_po_summary_ignores_item_order() {
-        let first = PoSemanticSummary::from_po_file(&PoFile {
-            headers: vec![ferrocat_po::Header {
-                key: "Language".to_owned(),
-                value: "de".to_owned(),
-            }],
-            items: vec![
-                ferrocat_po::PoItem {
-                    msgid: "b".to_owned(),
-                    msgstr: MsgStr::Singular("B".to_owned()),
-                    ..ferrocat_po::PoItem::default()
-                },
-                ferrocat_po::PoItem {
-                    msgid: "a".to_owned(),
-                    msgstr: MsgStr::Singular("A".to_owned()),
-                    ..ferrocat_po::PoItem::default()
-                },
-            ],
-            ..PoFile::default()
-        });
-        let second = PoSemanticSummary::from_po_file(&PoFile {
-            headers: vec![ferrocat_po::Header {
-                key: "Language".to_owned(),
-                value: "de".to_owned(),
-            }],
-            items: vec![
-                ferrocat_po::PoItem {
-                    msgid: "a".to_owned(),
-                    msgstr: MsgStr::Singular("A".to_owned()),
-                    ..ferrocat_po::PoItem::default()
-                },
-                ferrocat_po::PoItem {
-                    msgid: "b".to_owned(),
-                    msgstr: MsgStr::Singular("B".to_owned()),
-                    ..ferrocat_po::PoItem::default()
-                },
-            ],
-            ..PoFile::default()
-        });
-
-        assert_eq!(
-            digest_summary(&first).expect("digest"),
-            digest_summary(&second).expect("digest")
-        );
-    }
-
-    #[test]
-    fn statistics_use_elapsed_distribution() {
-        let stats = ScenarioStatistics::from_samples(&[
-            ExecutionResult {
-                tool_version: "tool".to_owned(),
-                reported_digest: "a".to_owned(),
-                elapsed_ns: 10,
-                baseline_elapsed_ns: None,
-                bytes_processed: 1024,
-                items_processed: Some(10),
-                messages_processed: None,
-                artifact: None,
-            },
-            ExecutionResult {
-                tool_version: "tool".to_owned(),
-                reported_digest: "a".to_owned(),
-                elapsed_ns: 30,
-                baseline_elapsed_ns: None,
-                bytes_processed: 1024,
-                items_processed: Some(10),
-                messages_processed: None,
-                artifact: None,
-            },
-            ExecutionResult {
-                tool_version: "tool".to_owned(),
-                reported_digest: "a".to_owned(),
-                elapsed_ns: 20,
-                baseline_elapsed_ns: None,
-                bytes_processed: 1024,
-                items_processed: Some(10),
-                messages_processed: None,
-                artifact: None,
-            },
-        ]);
-
-        assert_eq!(stats.median_elapsed_ns, 20);
-        assert_eq!(stats.min_elapsed_ns, 10);
-        assert_eq!(stats.max_elapsed_ns, 30);
-        assert_eq!(stats.median_absolute_deviation_ns, 10);
-        assert!(stats.mean_elapsed_ns > 0.0);
-        assert!(stats.stddev_elapsed_ns > 0.0);
-        assert!(stats.coefficient_of_variation > 0.0);
-        assert!(stats.relative_span_percent > 0.0);
-        assert!(stats.noisy);
-    }
-
-    #[test]
-    fn regression_check_passes_growth_within_threshold() {
-        let baseline = regression_report([regression_scenario_report(
-            "po-parse/mixed-10000/ferrocat-owned",
-            100_000_000,
-            false,
-        )]);
-        let current = regression_report([regression_scenario_report(
-            "po-parse/mixed-10000/ferrocat-owned",
-            110_000_000,
-            false,
-        )]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(!report.has_failures());
-        assert_eq!(report.passed.len(), 1);
-        assert!(report.render().contains("result: PASS"));
-    }
-
-    #[test]
-    fn regression_check_flags_meaningful_slowdown() {
-        let baseline = regression_report([regression_scenario_report(
-            "po-update/catalog-icu-heavy/ferrocat",
-            100_000_000,
-            false,
-        )]);
-        let current = regression_report([regression_scenario_report(
-            "po-update/catalog-icu-heavy/ferrocat",
-            130_000_000,
-            false,
-        )]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(report.has_failures());
-        assert_eq!(report.failures.len(), 1);
-        assert!(
-            report
-                .render()
-                .contains("FAIL po-update/catalog-icu-heavy/ferrocat")
-        );
-    }
-
-    #[test]
-    fn regression_check_skips_noisy_scenarios() {
-        let baseline = regression_report([regression_scenario_report(
-            "icu-parse/icu-nested-1000/ferrocat",
-            100_000_000,
-            false,
-        )]);
-        let current = regression_report([regression_scenario_report(
-            "icu-parse/icu-nested-1000/ferrocat",
-            180_000_000,
-            true,
-        )]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(!report.has_failures());
-        assert_eq!(report.skipped_noisy.len(), 1);
-        assert!(
-            report
-                .render()
-                .contains("SKIP noisy icu-parse/icu-nested-1000/ferrocat")
-        );
-    }
-
-    #[test]
-    fn regression_check_skips_semantic_digest_changes() {
-        let baseline = regression_report([regression_scenario_report_with_digest(
-            "po-parse/mixed-10000/ferrocat-owned",
-            100_000_000,
-            false,
-            "digest-a",
-        )]);
-        let current = regression_report([regression_scenario_report_with_digest(
-            "po-parse/mixed-10000/ferrocat-owned",
-            140_000_000,
-            false,
-            "digest-b",
-        )]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(!report.has_failures());
-        assert_eq!(report.skipped_semantics_changed.len(), 1);
-        assert!(
-            report
-                .render()
-                .contains("SKIP semantics-changed po-parse/mixed-10000/ferrocat-owned")
-        );
-    }
-
-    #[test]
-    fn regression_check_fails_when_baseline_scenario_disappears() {
-        let baseline = regression_report([regression_scenario_report(
-            "po-parse/mixed-10000/ferrocat-borrowed",
-            100_000_000,
-            false,
-        )]);
-        let current = regression_report([]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(report.has_failures());
-        assert_eq!(
-            report.missing_current,
-            vec!["po-parse/mixed-10000/ferrocat-borrowed"]
-        );
-        assert!(
-            report
-                .render()
-                .contains("FAIL missing-current po-parse/mixed-10000/ferrocat-borrowed")
-        );
-    }
-
-    #[test]
-    fn round_robin_schedule_interleaves_scenarios_by_round() {
-        assert_eq!(round_robin_schedule(&[2, 1, 3]), vec![0, 1, 2, 0, 2, 2]);
-    }
-
-    #[test]
-    fn calibrate_iterations_uses_median_probe_elapsed() {
-        let iterations = calibrate_iterations(250, &[10_000_000, 20_000_000, 90_000_000]);
-
-        assert_eq!(iterations, 13);
-    }
-
-    #[test]
-    fn profile_loads_serious_v1() {
-        let workspace = workspace_root().expect("workspace");
-        let profile = BenchmarkProfile::load(&workspace, "serious-v1").expect("profile");
-        assert_eq!(profile.name, "serious-v1");
-        assert!(!profile.scenarios.is_empty());
-    }
-
-    #[test]
-    fn profile_loads_gettext_compat_v1() {
-        let workspace = workspace_root().expect("workspace");
-        let profile = BenchmarkProfile::load(&workspace, "gettext-compat-v1").expect("profile");
-        assert_eq!(profile.name, "gettext-compat-v1");
-        assert!(!profile.scenarios.is_empty());
-    }
-
-    #[test]
-    fn profile_loads_gettext_official_v1() {
-        let workspace = workspace_root().expect("workspace");
-        let profile = BenchmarkProfile::load(&workspace, "gettext-official-v1").expect("profile");
-        assert_eq!(profile.name, "gettext-official-v1");
-        assert!(!profile.scenarios.is_empty());
-    }
-
-    #[test]
-    fn profile_loads_gettext_official_quick_v1() {
-        let workspace = workspace_root().expect("workspace");
-        let profile =
-            BenchmarkProfile::load(&workspace, "gettext-official-quick-v1").expect("profile");
-        assert_eq!(profile.name, "gettext-official-quick-v1");
-        assert!(!profile.scenarios.is_empty());
-        assert_eq!(profile.tool_requirement(), ToolRequirement::External);
-    }
-
-    #[test]
-    fn profile_loads_rust_scheduled_v1_without_external_tools() {
-        let workspace = workspace_root().expect("workspace");
-        let profile = BenchmarkProfile::load(&workspace, "rust-scheduled-v1").expect("profile");
-        assert_eq!(profile.name, "rust-scheduled-v1");
-        assert!(!profile.scenarios.is_empty());
-        assert_eq!(profile.tool_requirement(), ToolRequirement::RustOnly);
-    }
-
-    #[test]
-    fn profile_loads_gettext_workflows_ecosystem_v1() {
-        let workspace = workspace_root().expect("workspace");
-        let profile =
-            BenchmarkProfile::load(&workspace, "gettext-workflows-ecosystem-v1").expect("profile");
-        assert_eq!(profile.name, "gettext-workflows-ecosystem-v1");
-        assert!(!profile.scenarios.is_empty());
-    }
-
-    #[test]
-    fn profile_loads_storage_formats_v1() {
-        let workspace = workspace_root().expect("workspace");
-        let profile = BenchmarkProfile::load(&workspace, "storage-formats-v1").expect("profile");
-        assert_eq!(profile.name, "storage-formats-v1");
-        assert!(!profile.scenarios.is_empty());
-    }
-
-    #[test]
-    fn adapter_response_schema_accepts_optional_artifacts() {
-        let response = serde_json::from_str::<AdapterResponse>(
-            r#"{
-                "implementation":"polib",
-                "workload":"po-parse",
-                "fixture":"mixed-1000",
-                "success":true,
-                "semantic_digest":"abc",
-                "elapsed_ns":123,
-                "bytes_processed":456,
-                "items_processed":10,
-                "messages_processed":null,
-                "tool_version":"polib 1.0",
-                "po_summary":{"headers":[],"items":[]},
-                "icu_summary":null,
-                "po_output_path":null
-            }"#,
-        )
-        .expect("schema");
-
-        assert_eq!(response.implementation, "polib");
-        assert!(response.po_summary.is_some());
-    }
-
-    #[test]
-    fn benchmark_environment_reports_missing_tools() {
-        let workspace = workspace_root().expect("workspace");
-        let error = BenchmarkEnvironment::detect(
-            &workspace,
-            Some(OsStr::new("/definitely-missing")),
-            ToolRequirement::External,
-        )
-        .expect_err("expected failure");
-        assert!(error.contains("benchmark environment verification failed"));
-    }
-
-    #[test]
-    fn run_profile_supports_internal_compare_groups() {
-        let workspace = workspace_root().expect("workspace");
-        let profile = BenchmarkProfile {
-            name: "test-internal".to_owned(),
-            minimum_sample_millis: 1,
-            scenarios: vec![
-                BenchmarkScenario {
-                    id: "po-parse/mixed-1000/owned".to_owned(),
-                    comparison_group: "po-parse/mixed-1000".to_owned(),
-                    workload: "po-parse".to_owned(),
-                    operation: "parse".to_owned(),
-                    fixture: "mixed-1000".to_owned(),
-                    implementation: "ferrocat-parse".to_owned(),
-                    warmup_runs: 1,
-                    measured_runs: 2,
-                    minimum_sample_millis: Some(1),
-                },
-                BenchmarkScenario {
-                    id: "po-parse/mixed-1000/borrowed".to_owned(),
-                    comparison_group: "po-parse/mixed-1000".to_owned(),
-                    workload: "po-parse".to_owned(),
-                    operation: "parse".to_owned(),
-                    fixture: "mixed-1000".to_owned(),
-                    implementation: "ferrocat-parse-borrowed".to_owned(),
-                    warmup_runs: 1,
-                    measured_runs: 2,
-                    minimum_sample_millis: Some(1),
-                },
-            ],
-        };
-        let environment = BenchmarkEnvironment {
-            git_sha: "test-sha".to_owned(),
-            system_label: "Test CPU (16 GB RAM, TestOS arm64)".to_owned(),
-            os: "test-os".to_owned(),
-            cpu_model: "test-cpu".to_owned(),
-            memory_bytes: 16 * 1024 * 1024 * 1024,
-            rustc_version: "rustc test".to_owned(),
-            node_version: "node test".to_owned(),
-            python_version: "python test".to_owned(),
-            msgmerge_version: "msgmerge test".to_owned(),
-            msgcat_version: "msgcat test".to_owned(),
-            node_adapter_version: "node adapters".to_owned(),
-            python_adapter_version: "python adapters".to_owned(),
-            python_program: PathBuf::from("python3"),
-        };
-
-        let report = run_profile(&workspace, &environment, &profile).expect("report");
-        assert_eq!(report.profile, "test-internal");
-        assert_eq!(report.scenarios.len(), 2);
-        assert_eq!(
-            report.scenarios[0].semantic_digest,
-            report.scenarios[1].semantic_digest
-        );
-    }
-
-    #[test]
-    fn format_memory_label_uses_clean_gigabyte_counts() {
-        assert_eq!(format_memory_label(32 * 1024 * 1024 * 1024), "32 GB RAM");
-    }
-
-    #[test]
-    fn build_system_label_combines_cpu_memory_and_os() {
-        assert_eq!(
-            build_system_label_with_os("Apple M1 Pro", 32 * 1024 * 1024 * 1024, "macOS arm64"),
-            "Apple M1 Pro (32 GB RAM, macOS arm64)"
-        );
-    }
-
-    #[test]
-    fn owned_and_borrowed_match_on_gettext_plural_fixture() {
-        let fixture =
-            crate::fixtures::fixture_by_name("gettext-commerce-pl-1000").expect("fixture");
-        let owned = ferrocat_po::parse_po(fixture.content()).expect("owned parse");
-        let borrowed = ferrocat_po::parse_po_borrowed(fixture.content()).expect("borrowed parse");
-
-        let owned_summary = PoSemanticSummary::from_po_file(&owned);
-        let borrowed_summary = PoSemanticSummary::from_borrowed_po_file(&borrowed);
-
-        assert_eq!(
-            canonical_json_string(&owned_summary).expect("owned json"),
-            canonical_json_string(&borrowed_summary).expect("borrowed json")
-        );
-    }
-
-    fn regression_report<const N: usize>(scenarios: [ScenarioReport; N]) -> CompareReport {
-        CompareReport {
-            profile: "rust-scheduled-v1".to_owned(),
-            generated_at: "2026-06-23T00:00:00Z".to_owned(),
-            reference_host_policy: "test".to_owned(),
-            environment: EnvironmentMetadata {
-                git_sha: "test".to_owned(),
-                system_label: "test".to_owned(),
-                os: "test".to_owned(),
-                cpu_model: "test".to_owned(),
-                memory_bytes: 1,
-                rustc_version: "rustc test".to_owned(),
-                node_version: "node test".to_owned(),
-                python_version: "python test".to_owned(),
-                msgmerge_version: "msgmerge test".to_owned(),
-                msgcat_version: "msgcat test".to_owned(),
-                node_adapter_version: "node adapters test".to_owned(),
-                python_adapter_version: "python adapters test".to_owned(),
-            },
-            scenarios: scenarios.into_iter().collect(),
-        }
-    }
-
-    fn regression_scenario_report(
-        id: &str,
-        median_elapsed_ns: u128,
-        noisy: bool,
-    ) -> ScenarioReport {
-        regression_scenario_report_with_digest(id, median_elapsed_ns, noisy, "digest")
-    }
-
-    fn regression_scenario_report_with_digest(
-        id: &str,
-        median_elapsed_ns: u128,
-        noisy: bool,
-        semantic_digest: &str,
-    ) -> ScenarioReport {
-        ScenarioReport {
-            id: id.to_owned(),
-            comparison_group: id
-                .rsplit_once('/')
-                .map_or(id, |(group, _)| group)
-                .to_owned(),
-            workload: "test".to_owned(),
-            operation: "test".to_owned(),
-            fixture: "test".to_owned(),
-            implementation: "ferrocat-test".to_owned(),
-            tool_version: "ferrocat@test".to_owned(),
-            iterations_per_sample: 1,
-            warmup_runs: 0,
-            measured_runs: 3,
-            semantic_digest: semantic_digest.to_owned(),
-            baseline_strategy: None,
-            baseline_fixture: None,
-            statistics: regression_statistics(median_elapsed_ns, noisy),
-            samples: Vec::new(),
-        }
-    }
-
-    fn regression_statistics(median_elapsed_ns: u128, noisy: bool) -> ScenarioStatistics {
-        ScenarioStatistics {
-            median_elapsed_ns,
-            mean_elapsed_ns: f64_from_u128(median_elapsed_ns),
-            min_elapsed_ns: median_elapsed_ns,
-            max_elapsed_ns: median_elapsed_ns,
-            stddev_elapsed_ns: 0.0,
-            median_absolute_deviation_ns: 0,
-            coefficient_of_variation: if noisy { 0.06 } else { 0.01 },
-            relative_span_percent: if noisy { 11.0 } else { 1.0 },
-            noisy,
-            median_mib_per_sec: 1.0,
-            median_units_per_sec: 1.0,
-            median_baseline_elapsed_ns: None,
-            median_adjusted_elapsed_ns: None,
-            median_adjusted_mib_per_sec: None,
-            median_adjusted_units_per_sec: None,
-        }
-    }
-
-    fn regression_scenario_report_with_iterations(
-        id: &str,
-        median_elapsed_ns: u128,
-        iterations_per_sample: usize,
-    ) -> ScenarioReport {
-        ScenarioReport {
-            iterations_per_sample,
-            ..regression_scenario_report(id, median_elapsed_ns, false)
-        }
-    }
-
-    #[test]
-    fn regression_check_normalizes_by_iterations_per_sample() {
-        // Calibration guard (PR #155): same work (matching digest), but the
-        // harness picked 29 iterations/sample for the baseline and 57 for the
-        // current run. Each operation got faster, so the raw median_elapsed_ns
-        // is larger while throughput improved. The check must read an
-        // improvement, not a regression.
-        let baseline = regression_report([regression_scenario_report_with_iterations(
-            "icu-parse/icu-nested-1000/ferrocat",
-            110_429_000,
-            29,
-        )]);
-        let current = regression_report([regression_scenario_report_with_iterations(
-            "icu-parse/icu-nested-1000/ferrocat",
-            142_051_000,
-            57,
-        )]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(
-            !report.has_failures(),
-            "calibration-only change must not fail: {}",
-            report.render()
-        );
-        // per iteration: 3.808 ms -> 2.492 ms, about a 35% improvement
-        assert_eq!(report.passed.len(), 1);
-        assert!(
-            report.passed[0].regression_percent < -30.0,
-            "expected improvement near -35%, got {:.2}%",
-            report.passed[0].regression_percent
-        );
-    }
-
-    #[test]
-    fn regression_check_flags_per_iteration_slowdown_across_calibrations() {
-        // The flip side: a genuine per-operation slowdown must still trip the
-        // gate even when calibration hides it in the raw sample time. Baseline
-        // 2 ms/iter (10 iters), current 3 ms/iter (5 iters) -> +50% per
-        // iteration, although the raw sample time shrank from 20 ms to 15 ms and
-        // the old raw comparison would have waved it through.
-        let baseline = regression_report([regression_scenario_report_with_iterations(
-            "icu-parse/icu-nested-1000/ferrocat",
-            20_000_000,
-            10,
-        )]);
-        let current = regression_report([regression_scenario_report_with_iterations(
-            "icu-parse/icu-nested-1000/ferrocat",
-            15_000_000,
-            5,
-        )]);
-
-        let report =
-            compare_regression_reports(&baseline, &current, 20.0).expect("regression report");
-
-        assert!(
-            report.has_failures(),
-            "per-iteration slowdown must fail: {}",
-            report.render()
-        );
-    }
-}
+include!("compare/tests.rs");
